% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.graphNEL}
\alias{as.graphNEL}
\title{Convert adjacency matrix to graphNEL object}
\usage{
as.graphNEL(amat)
}
\arguments{
\item{amat}{An adjacency matrix}
}
\value{
A \code{graphNEL} object, see  \code{\link[graph]{graphNEL-class}}.
}
\description{
Convert adjacency matrix to graphNEL object
}
