% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco-class.R
\name{as_disco}
\alias{as_disco}
\title{Wrap a Graph as a Disco Object}
\usage{
as_disco(graph, kn = knowledge(), class = "PDAG")
}
\arguments{
\item{graph}{A causal graph object}

\item{kn}{A \code{Knowledge} object. Default is an empty \code{Knowledge} object.}

\item{class}{A string describing the graph class.}
}
\value{
A \code{Disco} object containing a \link[caugi:caugi]{caugi::caugi} and a \code{Knowledge} object in a list.
}
\description{
Wraps a causal graph object (e.g. a \link[caugi:caugi]{caugi::caugi} graph, or a graph object
from \pkg{pcalg} or \pkg{bnlearn}) and a \code{Knowledge} object into a \code{Disco}
object, which is the output object of causal discovery methods used in
\pkg{causalDisco}.
}
\details{
The conversion from any graph type to a \link[caugi:caugi]{caugi::caugi} is handled by the \pkg{caugi}
package.
}
\seealso{
\code{\link[caugi:caugi]{caugi::caugi()}}

Other Extending causalDisco:
\code{\link[=distribute_engine_args]{distribute_engine_args()}},
\code{\link[=list_registered_engines]{list_registered_engines()}},
\code{\link[=list_registered_tetrad_algorithms]{list_registered_tetrad_algorithms()}},
\code{\link[=make_method]{make_method()}},
\code{\link[=make_runner]{make_runner()}},
\code{\link[=new_disco_method]{new_disco_method()}},
\code{\link[=register_engine]{register_engine()}},
\code{\link[=register_tetrad_algorithm]{register_tetrad_algorithm()}},
\code{\link[=reset_engine_registry]{reset_engine_registry()}},
\code{\link[=reset_tetrad_alg_registry]{reset_tetrad_alg_registry()}}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
