% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{autoplot.discography}
\alias{autoplot.discography}
\title{Autoplot for discography objects}
\usage{
\method{autoplot}{discography}(
  object,
  layout = "graphopt",
  arrow_length = grid::unit(4, "mm"),
  circle_size = 2.5,
  node_size = 5,
  label_size = 3.5,
  label_vjust = -0.8,
  ...
)
}
\arguments{
\item{object}{A discography object.}

\item{layout}{ggraph layout algorithm.}

\item{arrow_length}{Length of arrow heads.}

\item{circle_size}{Size of hollow‐circle markers.}

\item{node_size}{Size of node points.}

\item{label_size}{Size of node text labels.}

\item{label_vjust}{Vertical adjustment of node labels.}

\item{...}{Extra arguments passed to ggraph::create_layout().}
}
\value{
A ggplot object.
}
\description{
Autoplot for discography objects
}
