% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{average_degree}
\alias{average_degree}
\title{Compute average degree for adjacency matrix}
\usage{
average_degree(amat)
}
\arguments{
\item{amat}{An adjacency matrix}
}
\value{
A numeric.
}
\description{
Computes the average degree, i.e. the number of edges divided
by the number of nodes.
}
