% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnlearn_search.R
\name{bnlearnSearch}
\alias{bnlearnSearch}
\title{bnlearnSearch – a tidy R6 front-end for \strong{bnlearn}}
\value{
An R6 object with the methods documented below.
}
\description{
A wrapper that lets you drive \code{bnlearns}’s structure-learning
algorithms within the \code{causalDisco} framework.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-bnlearnSearch-new}{\code{bnlearnSearch$new()}}
\item \href{#method-bnlearnSearch-set_params}{\code{bnlearnSearch$set_params()}}
\item \href{#method-bnlearnSearch-set_data}{\code{bnlearnSearch$set_data()}}
\item \href{#method-bnlearnSearch-set_test}{\code{bnlearnSearch$set_test()}}
\item \href{#method-bnlearnSearch-set_score}{\code{bnlearnSearch$set_score()}}
\item \href{#method-bnlearnSearch-set_alg}{\code{bnlearnSearch$set_alg()}}
\item \href{#method-bnlearnSearch-set_knowledge}{\code{bnlearnSearch$set_knowledge()}}
\item \href{#method-bnlearnSearch-run_search}{\code{bnlearnSearch$run_search()}}
\item \href{#method-bnlearnSearch-clone}{\code{bnlearnSearch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bnlearnSearch-new"></a>}}
\if{latex}{\out{\hypertarget{method-bnlearnSearch-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bnlearnSearch$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bnlearnSearch-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-bnlearnSearch-set_params}{}}}
\subsection{Method \code{set_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bnlearnSearch$set_params(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of extra tuning parameters stored by \code{set_params()}
and spliced into the learner call.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bnlearnSearch-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-bnlearnSearch-set_data}{}}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bnlearnSearch$set_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \code{data.frame} holding the data set currently attached to the
search object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bnlearnSearch-set_test"></a>}}
\if{latex}{\out{\hypertarget{method-bnlearnSearch-set_test}{}}}
\subsection{Method \code{set_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bnlearnSearch$set_test(method, alpha = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bnlearnSearch-set_score"></a>}}
\if{latex}{\out{\hypertarget{method-bnlearnSearch-set_score}{}}}
\subsection{Method \code{set_score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bnlearnSearch$set_score(method)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bnlearnSearch-set_alg"></a>}}
\if{latex}{\out{\hypertarget{method-bnlearnSearch-set_alg}{}}}
\subsection{Method \code{set_alg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bnlearnSearch$set_alg(method, args = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bnlearnSearch-set_knowledge"></a>}}
\if{latex}{\out{\hypertarget{method-bnlearnSearch-set_knowledge}{}}}
\subsection{Method \code{set_knowledge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bnlearnSearch$set_knowledge(knowledge_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bnlearnSearch-run_search"></a>}}
\if{latex}{\out{\hypertarget{method-bnlearnSearch-run_search}{}}}
\subsection{Method \code{run_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bnlearnSearch$run_search(data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \code{data.frame} holding the data set currently attached to the
search object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bnlearnSearch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-bnlearnSearch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bnlearnSearch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
