% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_obj.R, R/java_r_utils.R
\name{cast_obj}
\alias{cast_obj}
\title{Cast a Java object to a superclass}
\usage{
cast_obj(obj)

cast_obj(obj)
}
\arguments{
\item{obj}{A Java object to be cast.}
}
\value{
A Java object cast to a superclass that Tetrad takes as input.

A Java object castedd to some superclass that Tetrad takes as input.
}
\description{
This function takes a Java object and casts it to a superclass, so it can be
used in the Tetrad Search class.
}
