% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDisco-package.R
\docType{package}
\name{causalDisco-package}
\alias{causalDisco}
\alias{causalDisco-package}
\title{Causal Disco package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Various tools for inferring causal models from observational data. The package includes an implementation of the temporal Peter-Clark (TPC) algorithm. Petersen, Osler and Ekstrøm (2021) \doi{10.1093/aje/kwab087}. It also includes general tools for evaluating differences in adjacency matrices, which can be used for evaluating performance of causal discovery procedures.
}
\section{System requirements}{

Cargo (Rust's package manager), rustc >= 1.80.0, xz.

Java (>= 21) for Tetrad functionality (optional; includes installer; see inst/java/LICENSE and inst/java/COPYRIGHT);
Source code: \url{https://github.com/cmu-phil/tetrad}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/annennenne/causalDisco}
  \item \url{https://bjarkehautop.github.io/causalDisco/}
  \item Report bugs at \url{https://github.com/annennenne/causalDisco/issues}
}

}
\author{
\strong{Maintainer}: Bjarke Hautop Kristensen \email{bjarke.kristensen@sund.ku.dk}

Authors:
\itemize{
  \item Anne Helby Petersen \email{ahpe@sund.ku.dk}
  \item Frederik Fabricius-Bjerre \email{ffb@di.ku.dk}
}

Other contributors:
\itemize{
  \item Tobias Ellegaard Larsen \email{tobias.ellegaard@sund.ku.dk} [contributor]
  \item Claus Thorn Ekstrøm \email{ekstrom@sund.ku.dk} [contributor]
}

}
