% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args_and_distribute_args}
\alias{check_args_and_distribute_args}
\title{Check arguments and distribute them to the appropriate functions}
\usage{
check_args_and_distribute_args(
  search,
  args,
  engine,
  alg,
  test = NULL,
  score = NULL
)
}
\arguments{
\item{search}{An object of class \code{TetradSearch}
(required if engine is "tetrad").}

\item{args}{A named list of arguments to be checked and distributed.}

\item{engine}{The engine to be used
(e.g., "tetrad", "pcalg", "bnlearn", "causalDisco").}

\item{alg}{The algorithm to be used
(e.g., "pc", "fci", "ges", etc.).}

\item{test}{The independence test to be used (if applicable).}

\item{score}{The scoring method to be used (if applicable).}
}
\value{
A list containing the distributed arguments for the algorithm, test,
and score.
}
\description{
This function checks if the provided arguments are valid for the specified
engine and algorithm, and distributes them accordingly.
}
\keyword{internal}
