% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args_and_distribute_args_causalDisco}
\alias{check_args_and_distribute_args_causalDisco}
\title{Check arguments and distribute them to causalDisco class functions}
\usage{
check_args_and_distribute_args_causalDisco(
  args,
  alg,
  test = NULL,
  score = NULL
)
}
\description{
Check arguments and distribute them to causalDisco class functions
}
\keyword{internal}
