% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args_and_distribute_args_tetrad}
\alias{check_args_and_distribute_args_tetrad}
\title{Check arguments and distribute them to TetradSearch class functions}
\usage{
check_args_and_distribute_args_tetrad(
  search,
  args,
  alg,
  test = NULL,
  score = NULL
)
}
\description{
Check arguments and distribute them to TetradSearch class functions
}
\keyword{internal}
