% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{check_knowledge_obj}
\alias{check_knowledge_obj}
\title{Verify that an object is a knowledge}
\usage{
check_knowledge_obj(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check that the object is a \code{knowledge} object. Mostly
for internal use in causalDisco.
}
\keyword{internal}
