% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two tpdag or tskeleton objects}
\usage{
compare(x, y = NULL)
}
\arguments{
\item{x}{First object}

\item{y}{Second object (optional)}
}
\description{
Compare edges in two tpdag objects or two tskeleton objects.
Note that they should be based on the same variables.
Only edge absence/presence is compared, not edge orientation.
}
