% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTest.R
\name{corTest}
\alias{corTest}
\title{Test for vanishing partial correlations}
\usage{
corTest(x, y, S, suffStat)
}
\arguments{
\item{x}{Index of x variable}

\item{y}{Index of y variable}

\item{S}{Index of S variable(s), possibly NULL}

\item{suffStat}{Sufficient statistic; list with data,
binary variables and order.}
}
\description{
Test for vanishing partial correlations
}
