% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init_helpers.R
\name{current_heap_gb}
\alias{current_heap_gb}
\title{Get current Java heap size in gigabytes}
\usage{
current_heap_gb()
}
\value{
A numeric value representing the current heap size in gigabytes.
}
\description{
This function retrieves the current Java heap size in gigabytes (based on
rough calculation).
}
