% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init_helpers.R
\name{default_heap}
\alias{default_heap}
\title{Default heap size for Java Virtual Machine}
\usage{
default_heap()
}
\description{
This function retrieves the default heap size for the Java Virtual Machine.
The standard size is 2 gigabytes.
}
\keyword{internal}
