% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco.R
\name{disco}
\alias{disco}
\title{disco function}
\usage{
disco(data, method, knowledge = NULL)
}
\arguments{
\item{data}{a data frame}

\item{method}{an algorithm/method object}

\item{knowledge}{a knowledge object}
}
\description{
disco function
}
