% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco-doc-helpers.R
\name{disco_algs_return_doc_pdag}
\alias{disco_algs_return_doc_pdag}
\title{Causal Discovery Algorithm Return Value PDAG}
\description{
Causal Discovery Algorithm Return Value PDAG
}
\section{Value}{

A function that takes a single argument \code{data} (a data frame). When called,
this function returns a list containing:
\itemize{
\item \code{knowledge} A \code{Knowledge} object with the background knowledge
used in the causal discovery algorithm. See \code{\link[=knowledge]{knowledge()}} for how to construct it.
\item \code{caugi} A \code{\link[caugi:caugi]{caugi::caugi}} object (of class \code{PDAG}) representing the learned causal graph
from the causal discovery algorithm.
}
}

\keyword{internal}
