% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco-doc-helpers.R
\name{disco_algs_return_doc_rfci_pag}
\alias{disco_algs_return_doc_rfci_pag}
\title{Causal Discovery Algorithm Return Value RFCI-PAG}
\description{
Causal Discovery Algorithm Return Value RFCI-PAG
}
\section{Value}{

A function that takes a single argument \code{data} (a data frame). When called,
this function returns a list containing:
\itemize{
\item \code{knowledge} A \code{Knowledge} object with the background knowledge
used in the causal discovery algorithm. See \code{\link[=knowledge]{knowledge()}} for how to construct it.
\item \code{caugi} A \code{\link[caugi:caugi]{caugi::caugi}} object representing the learned causal graph.
This graph is an RFCI-PAG (RFCI Partial Ancestral Graph), but since RFCI-PAGs are not yet
natively supported in \pkg{caugi}, it is currently stored with class \code{UNKNOWN}.
}
Please see the definition 3.2 of the paper referenced for definition of an RFCI-PAG,
and it's differences from a standard PAG.
}

\keyword{internal}
