% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco_method.R
\name{disco_method}
\alias{disco_method}
\title{Wrap a runner list into a “disco_method” closure}
\usage{
disco_method(builder, method_class)
}
\description{
A \code{disco_method} is a closure that wraps a builder function.
The builder function is expected to return a runner object.
The closure can be called with a dataframe, and it will
build a fresh runner with the empty knowledge. Use \code{set_knowledge}
to set knowledge into the method.

Every \code{disco_method} can be used in combination with \code{disco}. If you want
to build your own method, you can use this function to create a closure
that will run with \code{disco}.
}
\keyword{internal}
