% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco-doc-helpers.R
\name{disco_note}
\alias{disco_note}
\title{Disco Recommendation Note}
\description{
Disco Recommendation Note
}
\section{Recommendation}{

While it is possible to call the function returned directly with a data frame,
we recommend using \code{\link[=disco]{disco()}}. This provides a consistent interface and handles knowledge
integration.
}

\keyword{internal}
