% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.default}
\alias{discography.default}
\title{Default method for discography}
\usage{
\method{discography}{default}(x, ...)
}
\arguments{
\item{x}{An unsupported object.}

\item{...}{Unused, passed for extensibility.}
}
\value{
Never returns; errors.
}
\description{
If the input is not recognized, throw error.
}
