% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.EssGraph}
\alias{discography.EssGraph}
\title{#' Convert EssGraph object (used by pcalg::ges) to a discography object.}
\usage{
\method{discography}{EssGraph}(x, nodes = x$.nodes, ...)
}
\description{
#' Convert EssGraph object (used by pcalg::ges) to a discography object.
}
