% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.fciAlgo}
\alias{discography.fciAlgo}
\title{Convert fciAlgo object to a discography object.}
\usage{
\method{discography}{fciAlgo}(x, nodes = rownames(x@amat), ...)
}
\description{
Convert fciAlgo object to a discography object.
}
