% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.gAlgo}
\alias{discography.gAlgo}
\title{Convert gAlgo object to a discography object.}
\usage{
\method{discography}{gAlgo}(x, nodes = x@graph@nodes, ...)
}
\description{
Convert gAlgo object to a discography object.
}
