% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.graphNEL}
\alias{discography.graphNEL}
\title{Convert graphNEL object to a discography object.}
\usage{
\method{discography}{graphNEL}(x, nodes = graph::nodes(x), ...)
}
\description{
Convert graphNEL object to a discography object.
}
