% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.pcAlgo}
\alias{discography.pcAlgo}
\title{Convert pcAlgo object to a discography object.}
\usage{
\method{discography}{pcAlgo}(x, nodes = x@graph@nodes, ...)
}
\description{
Convert pcAlgo object to a discography object.
}
