% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.tamat}
\alias{discography.tamat}
\title{Convert a temporal adjacency matrix (tamat) to a discography tibble}
\usage{
\method{discography}{tamat}(x, nodes = NULL, ...)
}
\arguments{
\item{x}{A causal-graph object (see methods below).}

\item{nodes}{Optional character vector giving the desired node order.
If \code{NULL}, the unique node names are sorted alphabetically.}

\item{...}{Passed on to methods.}
}
\value{
A "discography" tibble.
}
\description{
Convert a temporal adjacency matrix (tamat) to a discography tibble
}
