% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.tetrad_graph}
\alias{discography.tetrad_graph}
\title{Convert tetrad_graph object to a discography object.}
\usage{
\method{discography}{tetrad_graph}(x, nodes = x$nodes, ...)
}
\description{
Convert tetrad_graph object to a discography object.
}
