% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-causaldisco-algs.R
\name{distribute_engine_args}
\alias{distribute_engine_args}
\title{Distribute and Validate Engine Arguments}
\usage{
distribute_engine_args(search, args, engine, alg)
}
\arguments{
\item{search}{R6 object, either \code{TetradSearch}, \code{BnlearnSearch}, \code{PcalgSearch}, or \code{CausalDiscoSearch}.}

\item{args}{List of arguments to distribute}

\item{engine}{Engine identifier, either "tetrad", "bnlearn", "pcalg", or "causalDisco"}

\item{alg}{Algorithm name}
}
\description{
This function checks the provided arguments against the expected arguments for the specified engine and algorithm,
and distributes them appropriately to the search object. It ensures that the arguments are valid for the given
engine and algorithm, and then sets them on the search object.
}
\seealso{
Other Extending causalDisco: 
\code{\link{new_disco_method}()},
\code{\link{register_tetrad_algorithm}()},
\code{\link{reset_tetrad_alg_registry}()}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
