% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_testing.R
\name{.add_edges}
\alias{.add_edges}
\title{Add one or many edges to a knowledge object}
\usage{
.add_edges(.kn, status, edge_type, from, to)
}
\description{
Add one or many edges to a knowledge object
}
\keyword{internal}
