% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init_helpers.R
\name{.check_if_pkgs_are_installed}
\alias{.check_if_pkgs_are_installed}
\title{Check if required packages are installed}
\usage{
.check_if_pkgs_are_installed(pkgs, function_name = NULL, class_name = NULL)
}
\arguments{
\item{pkgs}{Character vector of package names to check.}

\item{function_name}{Name of the function requiring the packages (string).}

\item{class_name}{Name of the R6 class requiring the packages (string).}
}
\value{
Invisibly returns TRUE if all packages are installed,
otherwise stops or warns.
}
\description{
Used in functions and classes in causalDisco to check if required packages
are installed with a more informative error message.
}
\keyword{internal}
