% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.default_tetrad_version}
\alias{.default_tetrad_version}
\title{Default Tetrad version used in the package}
\format{
An object of class \code{character} of length 1.
}
\usage{
.default_tetrad_version
}
\description{
Default Tetrad version used in the package
}
\keyword{internal}
