% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{.formula_vars}
\alias{.formula_vars}
\title{Extract variable names from the RHS of a \code{tier()} formula}
\usage{
.formula_vars(.kn, rhs)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{rhs}{A formula (e.g. \code{1 ~ V1 + V2}).}
}
\description{
Extract variable names from the RHS of a \code{tier()} formula
}
\examples{
### .formula_vars() example ###

# extract variable names from a RHS using tidyselect or symbols
data(tpcExample)
kn <- knowledge(tpcExample)

# tidyselect RHS
rhs1 <- rlang::expr(starts_with("youth"))
causalDisco:::.formula_vars(kn, rhs1)

# symbol arithmetic-style RHS (falls back to symbols)
rhs2 <- rlang::expr(child_x1 + youth_x3)
causalDisco:::.formula_vars(kn, rhs2)
}
\keyword{internal}
