% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcalg-type-tests.R
\name{.get_suff_stat}
\alias{.get_suff_stat}
\title{Build sufficient statistic for pcalg/micd/causalDisco tests}
\usage{
.get_suff_stat(X, method, adaptDF = TRUE, nlev = NULL)
}
\description{
Build sufficient statistic for pcalg/micd/causalDisco tests
}
\keyword{internal}
