% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{.new_knowledge}
\alias{.new_knowledge}
\title{Create a \code{knowledge} object}
\usage{
.new_knowledge(vars = character(), frozen = FALSE)
}
\arguments{
\item{vars}{Character vector of variable names.  Defaults to empty.}

\item{frozen}{Logical. If \code{TRUE}, no new variables can be added. Defaults to \code{FALSE}.}
}
\value{
An S3 object of class \code{"knowledge"}.
}
\description{
Create a \code{knowledge} object
}
\keyword{internal}
