% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init_helpers.R
\name{.read_line}
\alias{.read_line}
\title{Read a line from the console}
\usage{
.read_line(prompt)
}
\description{
Only written so we can mock for testing
}
\examples{
### .read_line() example ###

# Reads a line from the console
# (only run interactively)
if (interactive()) {
  name <- causalDisco:::.read_line("What's your name? ")
  cat("Hello,", name, "\n")
}
}
\keyword{internal}
