% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_testing.R
\name{.resolved_vars}
\alias{.resolved_vars}
\title{Resolve a tidy-select or character spec to character names}
\usage{
.resolved_vars(.kn, spec)
}
\description{
Resolve a tidy-select or character spec to character names
}
\keyword{internal}
