% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{.tier_row}
\alias{.tier_row}
\title{Find row in the tiers dataframe matching a given index or label}
\usage{
.tier_row(.kn, idx = NULL, label = NULL)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{idx}{A numeric index or \code{NULL}.}

\item{label}{A character label or \code{NULL}.}
}
\description{
Find row in the tiers dataframe matching a given index or label
}
