% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{.update_edge_tiers}
\alias{.update_edge_tiers}
\title{Recompute the \code{tier_from} and \code{tier_to} columns of the edges dataframe}
\usage{
.update_edge_tiers(.kn)
}
\description{
Used internally in \code{+.knowledge}.
}
\keyword{internal}
