% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi-helpers.R
\name{edges}
\alias{edges}
\title{Retrieve Edges}
\usage{
edges(cg)
}
\arguments{
\item{cg}{A \code{caugi} object.}
}
\value{
A tibble containing the edges.
}
\description{
This function retrieves the edges from a \code{caugi} object as a tibble.
}
\examples{
data(tpc_example)
cd_tges <- tges(engine = "causalDisco", score = "tbic")
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)
disco_cd_tges <- disco(data = tpc_example, method = cd_tges, knowledge = kn)
edges(disco_cd_tges$caugi)
}
