% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_registry.R
\docType{data}
\name{engine_registry}
\alias{engine_registry}
\title{Engine Registry for causalDisco}
\format{
An object of class \code{list} of length 4.
}
\usage{
engine_registry
}
\description{
This registry contains the supported engines for causalDisco.
}
\keyword{datasets}
