% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{essgraph_to_amat}
\alias{essgraph_to_amat}
\title{Convert essential graph to adjacency matrix}
\usage{
essgraph_to_amat(essgraph, p = length(essgraph$field(".nodes")))
}
\arguments{
\item{essgraph}{An \code{EssGraph} object}

\item{p}{The number of nodes in the graph}
}
\value{
An adjacency matrix (square matrix with 0/1 entries).
}
\description{
Extracts the adjacency matrix from an \code{\link[pcalg]{EssGraph-class}} object. This object is returned
by score-based causal discovery algorithms in the pcalg package.
}
