% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{F1}
\alias{F1}
\title{F1 score}
\usage{
F1(confusion)
}
\arguments{
\item{confusion}{Confusion matrix as obtained from \link{confusion}}
}
\value{
A numeric in [0,1].
}
\description{
Computes F1 score from a confusion matrix, see \link{confusion}.
The F1 score is defined as  \eqn{2 * TP/(2 * TP + FP + FN)}, where TP are true positives,
FP are false positives, and FN are false negatives. If TP + FP + FN = 0, 1 is returned.
}
