% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{false_omission_rate}
\alias{false_omission_rate}
\title{False Omission Rate}
\usage{
false_omission_rate(truth, est, type = c("adj", "dir"))
}
\arguments{
\item{truth}{A \link[caugi:caugi]{caugi::caugi} object representing the truth graph.}

\item{est}{A \link[caugi:caugi]{caugi::caugi} object representing the estimated graph.}

\item{type}{Character string specifying the comparison type:
\itemize{
\item \code{"adj"}: adjacency comparison.
\item \code{"dir"}: orientation comparison conditional on shared adjacencies.
}}
}
\value{
A numeric in [0,1].
}
\description{
Computes false omission rate from two PDAG \link[caugi:caugi]{caugi::caugi} objects.
It converts the \link[caugi:caugi]{caugi::caugi} objects to adjacency matrices and computes
false omission rate as \code{FN/(FN + TN)}, where \code{FN} are false negatives and
\code{TN} are true negatives. If \verb{FN + TN = 0, 1} is returned.

Only supports \link[caugi:caugi]{caugi::caugi} objects whose edges are restricted to
\verb{-->}, \verb{<->}, \verb{---}, or absence of an edge.
}
\examples{
cg1 <- caugi::caugi(A \%-->\% B + C)
cg2 <- caugi::caugi(B \%-->\% A + C)
false_omission_rate(cg1, cg2, type = "adj")
false_omission_rate(cg1, cg2, type = "dir")

}
\seealso{
Other metrics:
\code{\link[=confusion]{confusion()}},
\code{\link[=evaluate]{evaluate()}},
\code{\link[=f1_score]{f1_score()}},
\code{\link[=fdr]{fdr()}},
\code{\link[=g1_score]{g1_score()}},
\code{\link[=npv]{npv()}},
\code{\link[=precision]{precision()}},
\code{\link[=recall]{recall()}},
\code{\link{reexports}},
\code{\link[=specificity]{specificity()}}
}
\concept{metrics}
