% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fci.R, R/fci_method.R
\name{fci}
\alias{fci}
\title{Perform causal discovery using the FCI algorithm}
\usage{
fci(engine = c("tetrad", "pcalg", "bnlearn"), test, alpha = 0.05, ...)

fci(engine = c("tetrad", "pcalg", "bnlearn"), test, alpha = 0.05, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"tetrad"}}{Tetrad Java library.}
\item{\code{"pcalg"}}{\pkg{pcalg} R package.}
\item{\code{"bnlearn"}}{\pkg{bnlearn} R package.}
}}

\item{test}{Character; name of the conditional‐independence test.}

\item{alpha}{Numeric; significance level for the CI tests.}

\item{...}{Additional arguments passed to the chosen engine (e.g. test or algorithm parameters).}

\item{output}{One of \code{"pag"} or \code{"fciAlgo"}. If \code{"pag"}
a partial ancestral graph (PAG) object is outputted.
If \code{"fciAlgo"} the PAG is outputted as the
object class \code{\link[pcalg]{fciAlgo-class}} from the pcalg package. This is
intended for compatability with tools from that package.}
}
\value{
A function of class \code{"fci"} that takes a single argument \code{data}
(a data frame) and returns an igraph_party object.
}
\description{
This is a wrapper function for the \code{\link[pcalg]{fci}} function as
implemented in the pcalg package. All computations are carried out by the
pcalg package. The default output object however matches that of \link{tfci}, see
this function for details about how the adjacency matrix is encoded.

Run the FCI algorithm for causal discovery using one of several engines.
}
\examples{
# simulate linear Gaussian data w unobserved variable L1
n <- 100
L1 <- rnorm(n) 
X1 <- rnorm(n)
X2 <- L1 + X1 + rnorm(n)
X3 <- X1 + rnorm(n)
X4 <- X3 + L1 + rnorm(n)
d <- data.frame(p1_X1 = X1,
                p1_X2 = X2,
                p2_X3 = X3,
                p2_X4 = X4)

# use FCI algorithm to recover PAG                
fci(d, test = corTest)


}
