% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpc_run.R
\name{find_adjacencies}
\alias{find_adjacencies}
\title{Find adjacencies of a node in an adjacency matrix}
\usage{
find_adjacencies(amatrix, index)
}
\arguments{
\item{amatrix}{Square adjacency matrix.}

\item{index}{Integer index of the node.}
}
\value{
Integer vector of adjacent node indices.
}
\description{
Find adjacencies of a node in an adjacency matrix
}
\examples{
### find_adjacencies() example ###

# Small toy adjacency (from-to)
amat <- matrix(0, 3, 3,
  dimnames = list(c("A", "B", "C"), c("A", "B", "C"))
)
# A --> B, C --> B
amat["A", "B"] <- 1
amat["C", "B"] <- 1

# Adjacent to B (both incoming and outgoing considered)
causalDisco:::find_adjacencies(amat, 2L) # indices of A and C (1, 3)
}
\keyword{internal}
