% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init_helpers.R
\name{find_tetrad_jar}
\alias{find_tetrad_jar}
\title{Find Tetrad GUI Launcher JAR}
\usage{
find_tetrad_jar(
  version = getOption("causalDisco.tetrad.version"),
  dir = getOption("tetrad.dir", Sys.getenv("TETRAD_DIR", ""))
)
}
\arguments{
\item{version}{Character string specifying the Tetrad version to search for.
Defaults to the package option \code{causalDisco.tetrad.version}.
You can override this by setting, for example:
\code{options(causalDisco.tetrad.version = "7.6.8")} \strong{before loading the package}.}

\item{dir}{Character string specifying the directory to search. Defaults to the value of
\code{getOption("tetrad.dir", Sys.getenv("TETRAD_DIR", ""))}.}
}
\value{
A character vector of length 1 containing the path to the Tetrad GUI launcher JAR.
If the file is not found, a warning is issued and an empty vector is returned.
}
\description{
\code{find_tetrad_jar()} searches for the Tetrad GUI launcher JAR file for a
specific version in a given directory. By default, it looks in the directory
set by the \code{tetrad.dir} option or the \code{TETRAD_DIR} environment variable.
}
\examples{
\dontrun{
gui_jar <- find_tetrad_jar()
print(gui_jar)
}

}
\keyword{internal}
