% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init_helpers.R
\name{find_tetrad_jars}
\alias{find_tetrad_jars}
\title{Find Tetrad .jar files}
\usage{
find_tetrad_jars()
}
\value{
A character vector of file paths to the JAR files.
}
\description{
This function searches for Tetrad JAR files in the package's Java directory.
}
\examples{
### find_tetrad_jars() example ###
jars <- causalDisco:::find_tetrad_jars()
length(jars) # usually 1, if installed
print(jars) # path to the jar file
}
\keyword{internal}
