% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{FOR}
\alias{FOR}
\title{False Omission Rate}
\usage{
FOR(confusion)
}
\arguments{
\item{confusion}{Confusion matrix as obtained from \link{confusion}}
}
\value{
A numeric in [0,1].
}
\description{
Computes false omission rate from a confusion matrix, see \link{confusion}.
False omission rate is defined as FN/(FN + TN), where FN are false negatives and
TN are true negatives. If FN + TN = 0, 0 is returned.
}
