% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_dag_data}
\alias{generate_dag_data}
\title{Generate Synthetic Data from a Linear Gaussian DAG}
\usage{
generate_dag_data(
  cg,
  n,
  ...,
  standardize = TRUE,
  coef_range = c(0.1, 0.9),
  error_sd = c(0.3, 2),
  seed = NULL
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been renamed to \code{\link[=sim_data]{sim_data()}} instead.
}
\keyword{internal}
