% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge-verbs.R
\name{get_tiers}
\alias{get_tiers}
\title{Get Tiers from Knowledge}
\usage{
get_tiers(kn)
}
\arguments{
\item{kn}{A \code{Knowledge} object.}
}
\value{
A tibble with the tiers.
}
\description{
Get tiers from a \code{Knowledge} object.
}
\examples{
kn <- knowledge(
  tier(
    1 ~ V1 + V2,
    2 ~ V3
  )
)

get_tiers(kn)
}
\seealso{
Other knowledge helpers:
\code{\link[=+.Knowledge]{+.Knowledge()}},
\code{\link[=add_exogenous]{add_exogenous()}},
\code{\link[=add_tier]{add_tier()}},
\code{\link[=add_to_tier]{add_to_tier()}},
\code{\link[=add_vars]{add_vars()}},
\code{\link[=as_bnlearn_knowledge]{as_bnlearn_knowledge()}},
\code{\link[=as_pcalg_constraints]{as_pcalg_constraints()}},
\code{\link[=as_tetrad_knowledge]{as_tetrad_knowledge()}},
\code{\link[=convert_tiers_to_forbidden]{convert_tiers_to_forbidden()}},
\code{\link[=deparse_knowledge]{deparse_knowledge()}},
\code{\link[=forbid_edge]{forbid_edge()}},
\code{\link[=knowledge]{knowledge()}},
\code{\link[=knowledge_to_caugi]{knowledge_to_caugi()}},
\code{\link[=remove_edge]{remove_edge()}},
\code{\link[=remove_tiers]{remove_tiers()}},
\code{\link[=remove_vars]{remove_vars()}},
\code{\link[=reorder_tiers]{reorder_tiers()}},
\code{\link[=reposition_tier]{reposition_tier()}},
\code{\link[=require_edge]{require_edge()}},
\code{\link[=seq_tiers]{seq_tiers()}},
\code{\link[=set_knowledge]{set_knowledge()}},
\code{\link[=set_max_lag]{set_max_lag()}},
\code{\link[=unfreeze]{unfreeze()}}
}
\concept{knowledge helpers}
\concept{knowledge-helper}
