% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTempoMech.R
\name{getvar.character}
\alias{getvar.character}
\title{Get variables with a specific prefix (character method)}
\usage{
\method{getvar}{character}(x, prefix, sep = "_")
}
\arguments{
\item{x}{Character vector}

\item{prefix}{Prefix to match}

\item{sep}{Separator, default "_"}
}
\value{
Subset of x that matches the prefix
}
\description{
Get variables with a specific prefix (character method)
}
