% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTempoMech.R
\name{getvar.data.frame}
\alias{getvar.data.frame}
\title{Get variables with a specific prefix (data.frame method)}
\usage{
\method{getvar}{data.frame}(x, prefix, sep = "_")
}
\arguments{
\item{x}{A data.frame}

\item{prefix}{Prefix to match}

\item{sep}{Separator, default "_"}
}
\value{
Names of columns in x that match the prefix
}
\description{
Get variables with a specific prefix (data.frame method)
}
