% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTempoMech.R
\name{getvar}
\alias{getvar}
\alias{getvar.character}
\alias{getvar.data.frame}
\title{Extract variables that share a prefix}
\usage{
getvar(x, prefix, sep = "_")

\method{getvar}{character}(x, prefix, sep = "_")

\method{getvar}{data.frame}(x, prefix, sep = "_")
}
\arguments{
\item{x}{Object to search: character vector or data frame.}

\item{prefix}{Character prefix of interest.}

\item{sep}{Character separator between prefix and remainder of name.}
}
\value{
Character vector of matching variable names.
}
\description{
Generic for selecting variables whose names start with a given prefix.
}
\keyword{internal}
