% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc.R
\name{hc}
\alias{hc}
\title{HC Algorithm for Causal Discovery}
\usage{
hc(engine = c("bnlearn"), score, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"bnlearn"}}{\pkg{bnlearn} R package.}
}}

\item{score}{Character; name of the scoring function to use.}

\item{...}{Additional arguments passed to the chosen engine (e.g. score or algorithm parameters).}
}
\value{
A function of class \code{"hc"} that takes a single argument \code{data}
(a data frame) and returns a \code{caugi} (of class "DAG") and a \code{knowledge} object.
}
\description{
Run the HC (hill-climbing) algorithm for causal discovery using one of several engines.
}
\details{
For specific details on the supported scores and parameters for each engine, see:
\itemize{
\item \code{\link{BnlearnSearch}} for \pkg{bnlearn}.
}

Note that this function \strong{always returns a fully directed DAG}.
In cases where the data do not provide enough information to identify the direction
of some edges (i.e., edges that are part of a Markov equivalence class),
the algorithm will choose a direction arbitrarily.
Therefore, some edges in the returned DAG may not reflect the true causal direction.
}
\examples{
data(tpc_example)

# Recommended path using disco()
hc_bnlearn <- hc(engine = "bnlearn", score = "aic_g")
disco(tpc_example, hc_bnlearn)

# or using hc_bnlearn directly
hc_bnlearn(tpc_example)

#### With tier knowledge ####
kn <- knowledge(
  tpc_example,
  tier(
    child ~ tidyselect::starts_with("child"),
    youth ~ tidyselect::starts_with("youth"),
    oldage ~ tidyselect::starts_with("oldage")
  )
)

hc_bnlearn <- hc(engine = "bnlearn", score = "bge")
disco(tpc_example, hc_bnlearn, knowledge = kn)

# or using hc_bnlearn directly
hc_bnlearn <- hc_bnlearn |> set_knowledge(kn)
hc_bnlearn(tpc_example)

# With all algorithm arguments specified
hc_bnlearn <- hc(
  engine = "bnlearn",
  score = "loglik_g",
  debug = FALSE,
  restart = 10,
  perturb = 1,
  max.iter = Inf,
  maxp = Inf,
  optimized = TRUE
)
disco(tpc_example, hc_bnlearn)
}
\seealso{
Other causal discovery algorithms: 
\code{\link{boss}()},
\code{\link{boss_fci}()},
\code{\link{fci}()},
\code{\link{ges}()},
\code{\link{gfci}()},
\code{\link{grasp}()},
\code{\link{grasp_fci}()},
\code{\link{gs}()},
\code{\link{iamb-family}},
\code{\link{pc}()},
\code{\link{sp_fci}()},
\code{\link{tabu}()},
\code{\link{tfci}()},
\code{\link{tges}()},
\code{\link{tpc}()}
}
\concept{causal discovery algorithms}
\concept{cd_algorithms}
