% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamb.R
\name{iamb}
\alias{iamb}
\title{IAMB Algorithm for Causal Discovery}
\usage{
iamb(engine = c("bnlearn"), test, alpha = 0.05, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"bnlearn"}}{\pkg{bnlearn} R package.}
}}

\item{test}{Character; name of the conditional‐independence test.}

\item{alpha}{Numeric; significance level for the CI tests.}

\item{...}{Additional arguments passed to the chosen engine (e.g. test or algorithm parameters).}
}
\value{
A function of class \code{"iamb"} that takes a single argument \code{data}
(a data frame) and returns a \code{caugi} (of class "PDAG") and a \code{knowledge} object.
}
\description{
Run the IAMB (Incremental Association) algorithm for causal discovery using one of several engines.
}
\details{
For specific details on the supported scores, tests, and parameters for each engine, see:
\itemize{
\item \code{\link{BnlearnSearch}} for \pkg{bnlearn}.
}
}
\examples{
data(tpc_example)

kn <- knowledge(
  tpc_example,
  starts_with("child") \%-->\% starts_with("youth")
)


# Recommended path using disco()
iamb_bnlearn <- iamb(engine = "bnlearn", test = "fisher_z", alpha = 0.05)
disco(tpc_example, iamb_bnlearn, knowledge = kn)

# or using iamb_bnlearn directly
iamb_bnlearn <- iamb_bnlearn |> set_knowledge(kn)
iamb_bnlearn(tpc_example)


# With all algorithm arguments specified
iamb_bnlearn <- iamb(
  engine = "bnlearn",
  test = "fisher_z",
  alpha = 0.05,
  max.sx = 2,
  debug = FALSE,
  undirected = TRUE
)

disco(tpc_example, iamb_bnlearn)
}
\seealso{
Other causal discovery algorithms: 
\code{\link{boss}()},
\code{\link{boss_fci}()},
\code{\link{fast_iamb}()},
\code{\link{fci}()},
\code{\link{ges}()},
\code{\link{gfci}()},
\code{\link{grasp}()},
\code{\link{grasp_fci}()},
\code{\link{gs}()},
\code{\link{pc}()},
\code{\link{sp_fci}()},
\code{\link{tfci}()},
\code{\link{tges}()},
\code{\link{tpc}()}
}
\concept{causal discovery algorithms}
\concept{cd_algorithms}
