% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init_helpers.R
\name{init_java}
\alias{init_java}
\title{Initialize Java Virtual Machine for causalDisco}
\usage{
init_java(heap = default_heap())
}
\arguments{
\item{heap}{A string specifying the heap size for the JVM. "2g" for 2
gigabytes.}
}
\description{
This function initializes the Java Virtual Machine (JVM) for the causalDisco
package. It sets the heap size and classpath based on the Tetrad JAR files
found. If the JVM is already initialized, it adds the JARs to the classpath.
}
\examples{
### init_java() example ###

# Only run in an interactive session and before rJava is already initialized
if (requireNamespace("rJava", quietly = TRUE) && !rJava::.jniInitialized) {
  init_java(heap = default_heap())
}
}
\keyword{internal}
