% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_cpdag}
\alias{is_cpdag}
\title{Check for CPDAG}
\usage{
is_cpdag(amat)
}
\arguments{
\item{amat}{An adjacency matrix}
}
\value{
A logical.
}
\description{
Check for CPDAG
}
\details{
Check: Is adjacency matrix proper CPDAG? See \code{\link[pcalg]{isValidGraph}} for
definition.
}
