% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init_helpers.R
\name{is_interactive}
\alias{is_interactive}
\title{Check if the session is interactive}
\usage{
is_interactive()
}
\value{
A logical value indicating whether the session is interactive.
}
\description{
This function checks if the current R session is interactive. Only written
for mocking.
}
\examples{
### is_interactive() example ###
causalDisco:::is_interactive()
}
\keyword{internal}
