% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{is_knowledge}
\alias{is_knowledge}
\title{Verify that an object is a knowledge}
\usage{
is_knowledge(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check that the object is a \code{knowledge} object. Mostly
for internal use in causalDisco.
}
\examples{
### is_knowledge() example ###

kn <- knowledge()
causalDisco:::is_knowledge(kn) # TRUE
try(causalDisco:::is_knowledge(list())) # throws error
}
\seealso{
Other knowledge functions: 
\code{\link{+.knowledge}()},
\code{\link{add_exogenous}()},
\code{\link{add_tier}()},
\code{\link{add_to_tier}()},
\code{\link{add_vars}()},
\code{\link{as_bnlearn_knowledge}()},
\code{\link{as_pcalg_constraints}()},
\code{\link{as_tetrad_knowledge}()},
\code{\link{deparse_knowledge}()},
\code{\link{forbid_edge}()},
\code{\link{forbid_tier_violations}()},
\code{\link{get_tiers}()},
\code{\link{knowledge}()},
\code{\link{remove_edges}()},
\code{\link{remove_tiers}()},
\code{\link{remove_vars}()},
\code{\link{reorder_tiers}()},
\code{\link{reposition_tier}()},
\code{\link{require_edge}()},
\code{\link{seq_tiers}()},
\code{\link{unfreeze}()}
}
\concept{knowledge}
\concept{knowledge functions}
\keyword{internal}
