% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeable_caugi.R
\name{is_knowledgeable_caugi}
\alias{is_knowledgeable_caugi}
\title{Is it a \code{knowledgeable_caugi}?}
\usage{
is_knowledgeable_caugi(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object is of class \code{knowledgeable_caugi}, \code{FALSE} otherwise.
}
\description{
Is it a \code{knowledgeable_caugi}?
}
