% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_pdag}
\alias{is_pdag}
\title{Check for PDAG}
\usage{
is_pdag(amat)
}
\arguments{
\item{amat}{An adjacency matrix}
}
\value{
A logical.
}
\description{
Check for PDAG
}
\details{
Check: Is adjacency matrix proper PDAG? See \code{\link[pcalg]{isValidGraph}} for
definition.
}
