% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeable-caugi.R
\name{knowledge.knowledgeable_caugi}
\alias{knowledge.knowledgeable_caugi}
\title{Extract Knowledge from a Knowledable Caugi}
\usage{
knowledge.knowledgeable_caugi(x)
}
\arguments{
\item{x}{A \code{knowledgeable_caugi} object.}
}
\value{
The nested \code{knowledge} object.
}
\description{
S3 method to extract the \code{knowledge} object from a \code{knowledgeable_caugi}.
}
