% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-causaldisco-algs.R
\name{list_registered_tetrad_algorithms}
\alias{list_registered_tetrad_algorithms}
\title{List Registered Tetrad Algorithms}
\usage{
list_registered_tetrad_algorithms()
}
\value{
Character vector of algorithm names.
}
\description{
Returns the names of all custom registered Tetrad algorithms.
}
\seealso{
Other Extending causalDisco: 
\code{\link{distribute_engine_args}()},
\code{\link{new_disco_method}()},
\code{\link{register_tetrad_algorithm}()},
\code{\link{reset_tetrad_alg_registry}()}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
