% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{make_method}
\alias{make_method}
\title{Create a Custom Causal Discovery Method}
\usage{
make_method(
  method_name,
  engine,
  engine_fns,
  test = NULL,
  alpha = NULL,
  score = NULL,
  graph_class,
  ...
)
}
\arguments{
\item{method_name}{Character. The name of the method to create.}

\item{engine}{Character. The engine to use. Must be one of the names of
\code{engine_fns}.}

\item{engine_fns}{Named list of functions. Each element corresponds to an
engine and is a function that implements the causal discovery algorithm.}

\item{test}{Optional. A test statistic to pass to the engine function.}

\item{alpha}{Optional. A significance level to pass to the engine function.}

\item{score}{Optional. A score to pass to the engine function.}

\item{graph_class}{Character. The graph class that this method produces.}

\item{...}{Additional arguments passed to the engine function.}
}
\value{
A \code{disco_method} object with attributes \code{engine} and \code{graph_class}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Constructs a new causal discovery method that can be used
with the \code{\link[=disco]{disco()}} framework. Users can provide an engine, engine-specific
functions, and optional test and alpha parameters.
}
\seealso{
Other Extending causalDisco: 
\code{\link{distribute_engine_args}()},
\code{\link{list_registered_tetrad_algorithms}()},
\code{\link{make_runner}()},
\code{\link{new_disco_method}()},
\code{\link{register_tetrad_algorithm}()},
\code{\link{reset_tetrad_alg_registry}()}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
