% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{maxnedges}
\alias{maxnedges}
\title{Compute maximal number of edges for graph}
\usage{
maxnedges(p)
}
\arguments{
\item{p}{The number of nodes in the graph}
}
\description{
Compute maximal number of edges for graph
}
