% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mix_data}
\alias{mix_data}
\title{Simulated Mixed Data}
\format{
A data.frame with 1000 rows and 5 variables.
\describe{
\item{X1}{Categorical, from \code{cat_data$X1}.}
\item{X2}{Categorical, from \code{cat_data$X2}.}
\item{X3}{Categorical, from \code{cat_data$X3}.}
\item{Z}{Numeric, same as \code{num_data$Z}.}
\item{Y}{Numeric, same as \code{num_data$Y}.}
}
}
\usage{
mix_data
}
\description{
A dataset combining continuous and categorical variables. The first three variables are replaced
with categorical versions from \code{cat_data}.
}
\details{
The R code used to generate this dataset is as follows:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{data(num_data)
data(cat_data)
mix_data <- num_data
mix_data$X1 <- cat_data$X1
mix_data$X2 <- cat_data$X2
mix_data$X3 <- cat_data$X3
}\if{html}{\out{</div>}}
}
\examples{
data(mix_data)
head(mix_data)

}
\seealso{
\link{num_data}, \link{cat_data}
}
\concept{dataset}
\keyword{datasets}
