% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_dags.R
\name{n_dags}
\alias{n_dags}
\title{Number of different DAGs}
\usage{
n_dags(p)
}
\arguments{
\item{p}{The number of nodes.}
}
\value{
A numeric.
}
\description{
Computes the number of different possible DAGs that can be constructed over
a given number of nodes.
}
