% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDAGs.R
\name{nDAGs}
\alias{nDAGs}
\title{Number of different DAGs}
\usage{
nDAGs(p)
}
\arguments{
\item{p}{The number of nodes.}
}
\value{
A numeric.
}
\description{
Computes the number of different possible DAGs that can be constructed over
a given number of nodes.
}
