% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nedges}
\alias{nedges}
\title{Number of edges in adjacency matrix}
\usage{
nedges(amat)
}
\arguments{
\item{amat}{An adjacency matrix}
}
\value{
A numeric (non-negative integer).
}
\description{
Counts the number of edges in an adjacency matrix.
}
