% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-causaldisco-algs.R
\name{new_disco_method}
\alias{new_disco_method}
\title{Create a causalDisco method}
\usage{
new_disco_method(builder, name, engine, graph_class)
}
\arguments{
\item{builder}{A function returning a runner}

\item{name}{Algorithm name}

\item{engine}{Engine identifier}

\item{graph_class}{Output graph class}
}
\value{
A function of class \code{"disco_method"} that takes a single argument
\code{data} (a data frame) and returns a \code{caugi} object.
}
\description{
Create a causalDisco method
}
\seealso{
Other Extending causalDisco: 
\code{\link{distribute_engine_args}()},
\code{\link{register_tetrad_algorithm}()},
\code{\link{reset_tetrad_alg_registry}()}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
