% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{new_discography}
\alias{new_discography}
\title{Minimal constructor for discography objects}
\usage{
new_discography(x)
}
\arguments{
\item{x}{A tibble (or data.frame) containing at least
\code{from}, \code{to}, and \code{edge_type} columns.}
}
\value{
A tibble with class \code{c("discography", "tbl_df", "tbl", "data.frame")}.
}
\description{
Minimal constructor for discography objects
}
