% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeable-caugi.R
\name{new_knowledgeable_caugi}
\alias{new_knowledgeable_caugi}
\title{Create a Knowledgeable Caugi Object}
\usage{
new_knowledgeable_caugi(cg, kn)
}
\arguments{
\item{cg}{A \code{caugi} object}

\item{kn}{A \code{knowledge} object}
}
\description{
Create a Knowledgeable Caugi Object
}
