% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{npv}
\alias{npv}
\title{Negative Predictive Value}
\usage{
npv(truth, est, type = c("adj", "dir"))
}
\arguments{
\item{truth}{A \link[caugi:caugi]{caugi::caugi} object representing the truth graph.}

\item{est}{A \link[caugi:caugi]{caugi::caugi} object representing the estimated graph.}

\item{type}{Character string specifying the comparison type:
\itemize{
\item \code{"adj"}: adjacency comparison.
\item \code{"dir"}: orientation comparison conditional on shared adjacencies.
}}
}
\value{
A numeric in [0,1].
}
\description{
Computes negative predictive value from two PDAG \link[caugi:caugi]{caugi::caugi} objects.
It converts the \link[caugi:caugi]{caugi::caugi} objects to adjacency matrices and computes
negative predictive value as \code{TN/(TN + FN)}, where \code{TN} are true negatives and
\code{FN} are false negatives. If \code{TN + FN = 0}, \code{1} is returned.

Only supports \link[caugi:caugi]{caugi::caugi} objects whose edges are restricted to
\verb{-->}, \verb{<->}, \verb{---}, or absence of an edge.
}
\examples{
cg1 <- caugi::caugi(A \%-->\% B + C)
cg2 <- caugi::caugi(B \%-->\% A + C)
npv(cg1, cg2, type = "adj")
npv(cg1, cg2, type = "dir")

}
\seealso{
Other metrics:
\code{\link[=confusion]{confusion()}},
\code{\link[=evaluate]{evaluate()}},
\code{\link[=f1_score]{f1_score()}},
\code{\link[=false_omission_rate]{false_omission_rate()}},
\code{\link[=fdr]{fdr()}},
\code{\link[=g1_score]{g1_score()}},
\code{\link[=precision]{precision()}},
\code{\link[=recall]{recall()}},
\code{\link{reexports}},
\code{\link[=specificity]{specificity()}}
}
\concept{metrics}
