% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init_helpers.R
\name{parse_heap_gb}
\alias{parse_heap_gb}
\title{Parse heap size string}
\usage{
parse_heap_gb(x)
}
\arguments{
\item{x}{A string representing the heap size, e.g., "4g" or "4096m".}
}
\value{
A numeric value representing the heap size in gigabytes.
}
\description{
This function parses a heap size string and returns the size in gigabytes.
}
\examples{
### parse_heap_gb() example ###
# everything will return 4
causalDisco:::parse_heap_gb("4g")
causalDisco:::parse_heap_gb("4G")
causalDisco:::parse_heap_gb("4096m")
causalDisco:::parse_heap_gb("4096MB")
}
\keyword{internal}
