% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pag}
\alias{plot.pag}
\title{Plot partial ancestral graph (PAG)}
\usage{
\method{plot}{pag}(x, ...)
}
\arguments{
\item{x}{pag object
to be plotted (as outputted from \code{\link{fci}}).}

\item{...}{Currently not in use.}
}
\value{
No return value, the function is called for its side-effects (plotting).
}
\description{
Plot partial ancestral graph (PAG)
}
\author{
This code is a modification of the fciAlgo plotting method implemented
in the pcalg package.
}
