% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tamat}
\alias{plot.tamat}
\title{Plot adjacency matrix with order information}
\usage{
\method{plot}{tamat}(x, ...)
}
\arguments{
\item{x}{tamat (temporal adjacency matrix) object to be plotted
(as outputted from \code{\link{tamat}}).}

\item{...}{Further plotting arguments passed along to \code{\link{plotTempoMech}}.}
}
\value{
No return value, the function is called for its side-effects (plotting).
}
\description{
Plot adjacency matrix with order information
}
