% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tpag}
\alias{plot.tpag}
\title{Plot temporal partial ancestral graph (TPAG)}
\usage{
\method{plot}{tpag}(x, ...)
}
\arguments{
\item{x}{tpag object
to be plotted (as outputted from \code{\link{tfci}}).}

\item{...}{Currently not in use.}
}
\value{
No return value, the function is called for its side-effects (plotting).
}
\description{
Plot temporal partial ancestral graph (TPAG)
}
\author{
This code is a modification of the fciAlgo plotting method implemented
in the pcalg package.
}
