% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tpdag}
\alias{plot.tpdag}
\title{Plot temporal partially directed acyclic graph (TPDAG)}
\usage{
plot.tpdag(x, ...)
}
\arguments{
\item{x}{tpdag object to be plotted (as outputted from \code{\link{tpc}}).}
}
\description{
blabla.
}
