% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tskeleton}
\alias{plot.tskeleton}
\title{Plot temporal skeleton}
\usage{
\method{plot}{tskeleton}(x, ...)
}
\arguments{
\item{x}{tskeleton (temporal skeleton) object to be plotted
(as outputted from \code{\link{tpc}}).}

\item{...}{Further plotting arguments passed along to \code{\link{plotTempoMech}}.}
}
\value{
No return value, the function is called for its side-effects (plotting).
}
\description{
Plot temporal skeleton
}
