% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tempo_mech.R
\name{plot_tempo_mech}
\alias{plot_tempo_mech}
\title{Plot temporal data generating mechanism}
\usage{
plot_tempo_mech(
  x,
  add_time_axis = TRUE,
  add_psi = TRUE,
  var_labels = NULL,
  period_labels = NULL,
  colors = NULL,
  ...
)
}
\arguments{
\item{x}{The tpdag/tskeleton or tamat to plot.}

\item{add_time_axis}{Logical indicating whether a time axis should be
added to the plot.}

\item{add_psi}{Logical indicating whether the sparsity level should be
added to the plot.}

\item{var_labels}{A named list of variable labels.}

\item{period_labels}{A character vector with labels for periods.}

\item{colors}{A character vector with colors to use for marking periods.
Should have at least as many elements as the numbers of periods.}

\item{...}{Additional arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\value{
No return value, the function is called for its side-effects (plotting).
}
\description{
Plots tpdag, tskeleton and tamat objects.
}
