% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeable-caugi.R
\name{print.knowledgeable_caugi}
\alias{print.knowledgeable_caugi}
\title{Print method for knowledgeable_caugi objects}
\usage{
\method{print}{knowledgeable_caugi}(x, compact = FALSE, ...)
}
\arguments{
\item{x}{A \code{knowledgeable_caugi} object.}

\item{compact}{Logical; if \code{TRUE}, prints a more compact representation.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for knowledgeable_caugi objects
}
\examples{
data(tpc_example)
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)
cd_tges <- tpc(engine = "causalDisco", test = "fisher_z")
disco_cd_tges <- disco(data = tpc_example, method = cd_tges, knowledge = kn)
print(disco_cd_tges)
print(disco_cd_tges, compact = TRUE)

}
