% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tetrad.R
\name{print.tetrad_check}
\alias{print.tetrad_check}
\title{Print Tetrad check result}
\usage{
\method{print}{tetrad_check}(x, ...)
}
\arguments{
\item{x}{A tetrad_check object.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for tetrad_check objects.
}
\examples{
causalDisco:::create_output(TRUE, "7.6.7", "Tetrad is installed.") |> print()
}
