% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{recall}
\alias{recall}
\title{Recall}
\usage{
recall(truth, guess, type = c("adj", "dir"))
}
\arguments{
\item{truth}{A \code{caugi} object representing the true graph.}

\item{guess}{A \code{caugi} object representing the estimated graph.}

\item{type}{Character string specifying the comparison type:
\itemize{
\item \code{"adj"}: adjacency comparison.
\item \code{"dir"}: orientation comparison conditional on shared adjacencies.
}}
}
\value{
A numeric in \verb{[0,1]}.
}
\description{
Computes recall from two \code{caugi} objects.
It converts the \code{caugi} objects to adjacency matrices and computes
recall as \code{TP/(TP + FN)}, where \code{TP} are true positives and
\code{FN} are false negatives. If \code{TP + FN = 0}, \code{1} is returned.
}
\examples{
cg1 <- caugi::caugi(A \%-->\% B + C)
cg2 <- caugi::caugi(B \%-->\% A + C)
recall(cg1, cg2, type = "adj")
recall(cg1, cg2, type = "dir")

}
\seealso{
Other metrics: 
\code{\link{confusion}()},
\code{\link{f1_score}()},
\code{\link{false_omission_rate}()},
\code{\link{fdr}()},
\code{\link{g1_score}()},
\code{\link{npv}()},
\code{\link{precision}()},
\code{\link{shd}()},
\code{\link{specificity}()}
}
\concept{metrics}
