% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-causaldisco-algs.R
\name{register_tetrad_algorithm}
\alias{register_tetrad_algorithm}
\title{Register a new Tetrad algorithm}
\usage{
register_tetrad_algorithm(name, setup_fun)
}
\arguments{
\item{name}{Algorithm name (string)}

\item{setup_fun}{A function that sets up the Tetrad search object for the
algorithm. It should take the same arguments as the runner function for the algorithm.}
}
\description{
Register a new Tetrad algorithm
}
\seealso{
Other Extending causalDisco: 
\code{\link{distribute_engine_args}()},
\code{\link{new_disco_method}()},
\code{\link{reset_tetrad_alg_registry}()}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
