% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{remove_edges}
\alias{remove_edges}
\title{Remove edges from a knowledge object}
\usage{
remove_edges(.kn, ...)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{...}{One or more two‐sided formulas, e.g. \code{A ~ B} or \code{starts_with("X") ~ Y}.}
}
\value{
The updated \code{knowledge} object.
}
\description{
Drop any directed edge(s) matching the two‐sided formulas you supply.
Errors if no edges matched.
}
