% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-causaldisco-algs.R
\name{reset_engine_registry}
\alias{reset_engine_registry}
\title{Reset the Engine Registry}
\usage{
reset_engine_registry()
}
\description{
Clears all custom registered engines.
}
\seealso{
Other Extending causalDisco:
\code{\link[=as_disco]{as_disco()}},
\code{\link[=distribute_engine_args]{distribute_engine_args()}},
\code{\link[=list_registered_engines]{list_registered_engines()}},
\code{\link[=list_registered_tetrad_algorithms]{list_registered_tetrad_algorithms()}},
\code{\link[=make_method]{make_method()}},
\code{\link[=make_runner]{make_runner()}},
\code{\link[=new_disco_method]{new_disco_method()}},
\code{\link[=register_engine]{register_engine()}},
\code{\link[=register_tetrad_algorithm]{register_tetrad_algorithm()}},
\code{\link[=reset_tetrad_alg_registry]{reset_tetrad_alg_registry()}}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
