% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-causaldisco-algs.R
\name{reset_tetrad_alg_registry}
\alias{reset_tetrad_alg_registry}
\title{Reset the Tetrad Algorithm Registry}
\usage{
reset_tetrad_alg_registry()
}
\description{
Clears all custom registered algorithms.
}
\seealso{
Other Extending causalDisco: 
\code{\link{distribute_engine_args}()},
\code{\link{list_registered_tetrad_algorithms}()},
\code{\link{new_disco_method}()},
\code{\link{register_tetrad_algorithm}()}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
