% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTempoMech.R
\name{sciNotation}
\alias{sciNotation}
\title{Convert numeric to scientific-notation expression}
\usage{
sciNotation(x, digits = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{digits}{Integer number of significant digits for the mantissa.}
}
\value{
An \code{expression} (or vector of expressions) for pretty printing.
}
\description{
Produces an \code{expression} representing a single value in scientific
notation, useful for annotation in base graphics functions.
}
\details{
from https://stat.ethz.ch/pipermail/r-help/2006-January/086034.html
}
\keyword{internal}
