% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{shd}
\alias{shd}
\title{Structural Hamming Distance}
\usage{
shd(truth, guess)
}
\arguments{
\item{truth}{A \code{caugi} object representing the true graph.}

\item{guess}{A \code{caugi} object representing the estimated graph.}
}
\value{
A numeric representing the SHD.
}
\description{
Computes the Structural Hamming Distance (SHD) between two \code{caugi} objects.
SHD is the number of edge additions, deletions, or orientation flips required to
transform one graph into the other. Adjacency errors are counted fully, while
orientation errors on shared adjacencies are counted as half an error, following
the standard convention for partially directed graphs.
}
\examples{
cg1 <- caugi::caugi(A \%-->\% B + C)
cg2 <- caugi::caugi(B \%-->\% A + C)
shd(cg1, cg2)

}
\seealso{
Other metrics: 
\code{\link{confusion}()},
\code{\link{f1_score}()},
\code{\link{false_omission_rate}()},
\code{\link{fdr}()},
\code{\link{g1_score}()},
\code{\link{npv}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{specificity}()}
}
\concept{metrics}
