% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(truth, est, type = c("adj", "dir"))
}
\arguments{
\item{truth}{A \link[caugi:caugi]{caugi::caugi} object representing the truth graph.}

\item{est}{A \link[caugi:caugi]{caugi::caugi} object representing the estimated graph.}

\item{type}{Character string specifying the comparison type:
\itemize{
\item \code{"adj"}: adjacency comparison.
\item \code{"dir"}: orientation comparison conditional on shared adjacencies.
}}
}
\value{
A numeric in [0,1].
}
\description{
Computes specificity from two PDAG \link[caugi:caugi]{caugi::caugi} objects.
It converts the \link[caugi:caugi]{caugi::caugi} objects to adjacency matrices and computes
specificity as \code{TN/(TN + FP)}, where \code{TN} are true negatives and
\code{FP} are false positives. If \code{TN + FP = 0}, \code{1} is returned.

Only supports \link[caugi:caugi]{caugi::caugi} objects whose edges are restricted to
\verb{-->}, \verb{<->}, \verb{---}, or absence of an edge.
}
\examples{
cg1 <- caugi::caugi(A \%-->\% B + C)
cg2 <- caugi::caugi(B \%-->\% A + C)
specificity(cg1, cg2, type = "adj")
specificity(cg1, cg2, type = "dir")

}
\seealso{
Other metrics:
\code{\link[=confusion]{confusion()}},
\code{\link[=evaluate]{evaluate()}},
\code{\link[=f1_score]{f1_score()}},
\code{\link[=false_omission_rate]{false_omission_rate()}},
\code{\link[=fdr]{fdr()}},
\code{\link[=g1_score]{g1_score()}},
\code{\link[=npv]{npv()}},
\code{\link[=precision]{precision()}},
\code{\link[=recall]{recall()}},
\code{\link{reexports}}
}
\concept{metrics}
