% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeable-caugi.R
\name{summary.knowledgeable_caugi}
\alias{summary.knowledgeable_caugi}
\title{Summary method for knowledgeable_caugi objects}
\usage{
\method{summary}{knowledgeable_caugi}(object, ...)
}
\arguments{
\item{object}{A \code{knowledgeable_caugi} object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the \code{knowledgeable_caugi} object.
}
\description{
Summary method for knowledgeable_caugi objects
}
\examples{
data(tpc_example)
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)
cd_tges <- tpc(engine = "causalDisco", test = "fisher_z")
disco_cd_tges <- disco(data = tpc_example, method = cd_tges, knowledge = kn)
summary(disco_cd_tges)

}
