% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamat.R
\name{tamat}
\alias{tamat}
\title{Make a temporal adjacency matrix}
\usage{
tamat(amat, order)
}
\arguments{
\item{amat}{Adjacency matrix. A square matrix of 0 or 1s. A
1 in the (i,j)th entry means that there is an edge from j to i.
Row names and column names should be identical and be the names of
the variables/nodes. Variable names should be prefixed with their
period, e.g. "child_x" for variable "x" at period "child"}

\item{order}{A character vector with the periods in their order.}
}
\description{
Make a temporal adjacency matrix
}
