% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrad_graph.R
\name{tetrad_graph}
\alias{tetrad_graph}
\title{Build a lightweight S3 wrapper around a Tetrad PAG string}
\usage{
tetrad_graph(x)
}
\arguments{
\item{x}{Character scalar in the form
\verb{"V1,V2,V3\\n0,2,0\\n3,0,1\\n0,3,0\\n"}}
}
\value{
An object of class <tetrad_graph>
}
\description{
Build a lightweight S3 wrapper around a Tetrad PAG string
}
\examples{
### tetrad_graph() example ###

# A tiny 3-node PAG in Tetrad’s CSV-like string form:
# rows/cols are A,B,C; entries use pcalg’s PAG codes (0,1,2,3)
pag_txt <- "A,B,C
0,2,0
3,0,1
0,3,0
"

g <- tetrad_graph(pag_txt)

# Inspect the parsed object
class(g)
g$nodes
g$amat
}
