% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tges}
\alias{tges}
\title{TGES Algorithm for Causal Discovery}
\usage{
tges(engine = c("causalDisco"), score, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"causalDisco"}}{causalDisco library.}
}}

\item{score}{Character; name of the scoring function to use.}

\item{...}{Additional arguments passed to the chosen engine
(e.g. test or algorithm parameters).}
}
\description{
Run the Temporal Greedy Equivalent Search algorithm for causal discovery using one of
several engines.
}
\details{
For specific details on the supported scores, see \link{CausalDiscoSearch}. For additional parameters
passed via \code{...}, see \code{\link[=tges_run]{tges_run()}}.
}
\section{Recommendation}{

While it is possible to call the function returned directly with a data frame,
we recommend using \code{\link[=disco]{disco()}}. This provides a consistent interface and handles knowledge
integration.
}

\section{Value}{

A function that takes a single argument \code{data} (a data frame). When called,
this function returns a list containing:
\itemize{
\item \code{knowledge} A \code{Knowledge} object with the background knowledge
used in the causal discovery algorithm. See \code{\link[=knowledge]{knowledge()}} for how to construct it.
\item \code{caugi} A \code{\link[caugi:caugi]{caugi::caugi}} object (of class \code{PDAG}) representing the learned causal graph
from the causal discovery algorithm.
}
}

\examples{
# Recommended route using disco:
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

my_tges <- tges(engine = "causalDisco", score = "tbic")

disco(tpc_example, my_tges, knowledge = kn)

# another way to run it

my_tges <- my_tges |>
  set_knowledge(kn)
my_tges(tpc_example)


# or you can run directly with tges_run()

data(tpc_example)

score_bic <- new(
  "TemporalBIC",
  data = tpc_example,
  nodes = colnames(tpc_example),
  knowledge = kn
)

res_bic <- tges_run(score_bic)
res_bic
}
\references{
Larsen TE, Ekstrøm CT, and Petersen AH. Score-Based Causal Discovery with Temporal
Background Information, 2025. \url{doi:10.48550/arXiv.2502.06232}.
}
\seealso{
Other causal discovery algorithms:
\code{\link[=boss]{boss()}},
\code{\link[=boss_fci]{boss_fci()}},
\code{\link[=fci]{fci()}},
\code{\link[=ges]{ges()}},
\code{\link[=gfci]{gfci()}},
\code{\link[=grasp]{grasp()}},
\code{\link[=grasp_fci]{grasp_fci()}},
\code{\link[=gs]{gs()}},
\code{\link{iamb-family}},
\code{\link[=pc]{pc()}},
\code{\link[=rfci]{rfci()}},
\code{\link[=sp_fci]{sp_fci()}},
\code{\link[=tfci]{tfci()}},
\code{\link[=tpc]{tpc()}}
}
\concept{causal discovery algorithms}
\concept{cd_algorithms}
