% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges-run.R
\name{tges_run}
\alias{tges_run}
\title{Run the TGES Algorithm for Causal Discovery}
\usage{
tges_run(score, verbose = FALSE)
}
\arguments{
\item{score}{tiered scoring object to be used.
At the moment only scores supported are
\itemize{
\item \link{TemporalBIC} and
\item \link{TemporalBDeu}.
}}

\item{verbose}{indicates whether debug output should be printed.}
}
\description{
Perform causal discovery using the temporal greedy equivalence search algorithm.
}
\section{Recommendation}{

While it is possible to call the function returned directly with a data frame,
we recommend using \code{\link[=disco]{disco()}}. This provides a consistent interface and handles knowledge
integration.
}

\section{Value}{

A function that takes a single argument \code{data} (a data frame). When called,
this function returns a list containing:
\itemize{
\item \code{knowledge} A \code{Knowledge} object with the background knowledge
used in the causal discovery algorithm. See \code{\link[=knowledge]{knowledge()}} for how to construct it.
\item \code{caugi} A \code{\link[caugi:caugi]{caugi::caugi}} object (of class \code{PDAG}) representing the learned causal graph
from the causal discovery algorithm.
}
}

\examples{
# Recommended route using disco:
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

my_tges <- tges(engine = "causalDisco", score = "tbic")

disco(tpc_example, my_tges, knowledge = kn)

# another way to run it

my_tges <- my_tges |>
  set_knowledge(kn)
my_tges(tpc_example)


# or you can run directly with tges_run()

data(tpc_example)

score_bic <- new(
  "TemporalBIC",
  data = tpc_example,
  nodes = colnames(tpc_example),
  knowledge = kn
)

res_bic <- tges_run(score_bic)
res_bic
}
\author{
Tobias Ellegaard Larsen
}
