% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGausFromDAG.R
\name{topo_order_mat}
\alias{topo_order_mat}
\title{Topological ordering of a DAG given its adjacency matrix}
\usage{
topo_order_mat(amat)
}
\arguments{
\item{amat}{An adjacency matrix.}
}
\value{
A vector of node indices in topological order.
}
\description{
Topological ordering of a DAG given its adjacency matrix
}
\keyword{internal}
