% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpc.R
\name{tpc}
\alias{tpc}
\title{The Temporal Peter-Clark (PC) algorithm for causal discovery}
\usage{
tpc(engine = c("causalDisco"), test, alpha = 0.05, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"causalDisco"}}{causalDisco library.}
}}

\item{test}{Character; name of the conditional‐independence test.}

\item{alpha}{Numeric; significance level for the CI tests.}

\item{...}{Additional arguments passed to the chosen engine
(e.g. test or algorithm parameters).}
}
\value{
A function of class \code{"tpc"} that takes a single argument \code{data}
(a data frame) and returns an \code{discography} object.
}
\description{
Run the temporal PC algorithm for causal discovery using causalDisco.
}
