% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpc_run.R
\name{v_orient_temporal}
\alias{v_orient_temporal}
\title{Temporally orient unshielded colliders}
\usage{
v_orient_temporal(amat, sepsets)
}
\arguments{
\item{amat}{Square adjacency matrix (from-to convention).}

\item{sepsets}{Separation sets as computed by \pkg{pcalg}.}
}
\value{
The updated adjacency matrix with additional arrowheads.
}
\description{
Given a CPDAG adjacency matrix and separation sets, orient v-structures
that do not contradict the current directions, respecting temporal tiering.
}
\examples{
### v_orient_temporal() example ###

data(tpcExample)
vnames <- names(tpcExample)

kn <- knowledge(
  tpcExample,
  tier(
    child ~ tidyselect::starts_with("child"),
    youth ~ tidyselect::starts_with("youth"),
    oldage ~ tidyselect::starts_with("oldage")
  )
)

ss_cor <- causalDisco:::make_suff_stat(tpcExample, type = "corTest")
wrapped_test <- causalDisco:::dirTest(causalDisco::corTest, vnames, kn)
pc_cons <- causalDisco:::.pcalg_constraints_from_knowledge(
  kn,
  labels = vnames,
  directed_as_undirected = TRUE
)

set.seed(1)
skel <- pcalg::skeleton(
  suffStat   = ss_cor,
  indepTest  = wrapped_test,
  alpha      = 0.05,
  labels     = vnames,
  method     = "stable.fast",
  fixedGaps  = pc_cons$fixedGaps,
  fixedEdges = pc_cons$fixedEdges
)

amat_skel <- causalDisco:::graph2amat(skel)
amat_prune <- causalDisco:::order_restrict_amat_cpdag(amat_skel, kn)
amat_orient <- causalDisco:::v_orient_temporal(amat_prune, skel@sepset)
amat_orient[1:6, 1:6]
}
\keyword{internal}
