% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{add_exogenous}
\alias{add_exogenous}
\alias{add_exo}
\alias{add_root}
\title{Add exogenous variables}
\usage{
add_exogenous(.kn, vars)

add_exo(.kn, vars)

add_root(.kn, vars)
}
\arguments{
\item{.kn}{A knowledge object.}

\item{vars}{Tidyselect specification or character vector of variables.}
}
\value{
Updated knowledge object.
}
\description{
Adds variables that cannot have incoming edges (root/exogenous nodes).
Every possible incoming edge to these nodes is automatically forbidden.
This is equivalent to writing \code{forbidden(everything() ~ vars)}.
}
