% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_testing.R
\name{add_tier}
\alias{add_tier}
\title{Assign variables to a tier (numeric or symbolic label)}
\usage{
add_tier(.kn, tier, vars)
}
\description{
Assign variables to a tier (numeric or symbolic label)
}
