% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{add_tier}
\alias{add_tier}
\title{Add (and position) a tier}
\usage{
add_tier(.kn, tier, before = NULL, after = NULL)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{tier}{Bare symbol / character (label) \strong{or} numeric literal.}

\item{before, after}{Optional anchor relative to an existing tier label,
tier index, or variable.  Once the knowledge object already
has >= 1 tier, you must supply \strong{exactly one} of these.}
}
\value{
Updated \code{knowledge} object.
}
\description{
Add (and position) a tier
}
