% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{add_to_tier}
\alias{add_to_tier}
\title{Add variables to an existing tier}
\usage{
add_to_tier(.kn, ...)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{...}{One or more two-sided formulas \code{tier ~ vars}.}
}
\value{
The updated \code{knowledge} object.
}
\description{
Add variables to an existing tier
}
