% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{add_vars}
\alias{add_vars}
\title{Add variables to \code{knowledge} object}
\usage{
add_vars(.kn, vars)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{vars}{A character vector of variable names to add.}
}
\description{
Adds variables to the \code{knowledge} object. If the object is
frozen, an error is thrown if any of the variables are not present in the
data frame provided to the object.
}
