% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{as_bnlearn_knowledge}
\alias{as_bnlearn_knowledge}
\title{Convert background knowledge to bnlearns white- and blacklists}
\usage{
as_bnlearn_knowledge(.kn)
}
\arguments{
\item{.kn}{A \code{knowledge} object.  Must have no tier information.}
}
\value{
A list with two elements, \code{whitelist} and \code{blacklist}, each a data
frame containing the edges in a \code{from}, \code{to} format.
}
\description{
Converts a \code{knowledge} object to a list of two data frames, namely
\code{whitelist} and \code{blacklist}, which can be used as arguments for
\code{bnlearn} algorithms. The \code{whitelist} contains all required edges, and the
\code{blacklist} containts all forbidden edges. Tiers will be made into forbidden
edges before running the conversion.
}
