% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{as_pcalg_constraints}
\alias{as_pcalg_constraints}
\title{Convert to a pair of (fixedGaps, fixedEdges) matrices for the \strong{pcalg} package}
\usage{
as_pcalg_constraints(.kn, labels)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{labels}{Character vector of variable names in the \emph{exact} order of
your data matrix / data frame.}
}
\value{
A list with elements \code{fixedGaps} and \code{fixedEdges}.
}
\description{
pcalg supports \emph{symmetric} background knowledge only:
• \strong{fixedGaps}  – \strong{forbidden} edges (undirected)
• \strong{fixedEdges} – \strong{required} edges (undirected)
}
\details{
We therefore throw an error if any edges are not undirected.
}
