% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{as_tetrad_knowledge}
\alias{as_tetrad_knowledge}
\title{Convert to Tetrad \code{edu.cmu.tetrad.data.Knowledge}}
\usage{
as_tetrad_knowledge(.kn)
}
\description{
Convert to Tetrad \code{edu.cmu.tetrad.data.Knowledge}
}
