% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{as_tetrad_knowledge}
\alias{as_tetrad_knowledge}
\title{Convert to Tetrad \code{edu.cmu.tetrad.data.Knowledge}}
\usage{
as_tetrad_knowledge(.kn)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}
}
\value{
A Java \code{edu.cmu.tetrad.data.Knowledge} object.
}
\description{
Converts a \code{knowledge} object to a Tetrad \code{edu.cmu.tetrad.data.Knowledge}.
This requires \code{rJava}. This is used internally, when setting knowledge with
\code{set_knowledge} for methods using the Tetrad engine. \code{set_knowledge} is used
internally, when using the \code{disco} function with knowledge given.
}
\examples{
# convert to Tetrad Knowledge via rJava
data(tpc_example)

kn <- knowledge(
  head(tpc_example),
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    oldage ~ starts_with("old")
  ),
  child_x1 \%-->\% youth_x3
)

jk <- try(as_tetrad_knowledge(kn)) # will run only if rJava/JVM available
try(print(jk)) # prints a Java reference if successful
}
\seealso{
Other knowledge functions: 
\code{\link{+.knowledge}()},
\code{\link{add_exogenous}()},
\code{\link{add_tier}()},
\code{\link{add_to_tier}()},
\code{\link{add_vars}()},
\code{\link{as_bnlearn_knowledge}()},
\code{\link{as_pcalg_constraints}()},
\code{\link{deparse_knowledge}()},
\code{\link{forbid_edge}()},
\code{\link{forbid_tier_violations}()},
\code{\link{get_tiers}()},
\code{\link{knowledge}()},
\code{\link{remove_edges}()},
\code{\link{remove_tiers}()},
\code{\link{remove_vars}()},
\code{\link{reorder_tiers}()},
\code{\link{reposition_tier}()},
\code{\link{require_edge}()},
\code{\link{seq_tiers}()},
\code{\link{unfreeze}()}
}
\concept{knowledge}
\concept{knowledge functions}
