% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_init_helpers.R
\name{ask_heap_size}
\alias{ask_heap_size}
\title{Ask for Java heap size}
\usage{
ask_heap_size()
}
\description{
This function prompts the user to specify the Java heap size if the heap
size is not set in the options or environment variables. Is only used in
interactive sessions.
}
