% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDisco-package.R
\name{causalDisco-package}
\alias{causalDisco-package}
\title{Causal Disco package}
\description{
Causal Disco package
}
\section{System requirements}{

Requires Java (version 17 or newer) to run the bundled Tetrad Java library.
You can check your Java version with \code{system("java -version")}.
}

