% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDisco-package.R
\docType{package}
\name{causalDisco-package}
\alias{causalDisco}
\alias{causalDisco-package}
\title{Causal Disco package}
\description{
Various tools for inferring causal models from observational data. The package includes an implementation of the temporal Peter-Clark (TPC) algorithm. Petersen, Osler and Ekstrøm (2021) \doi{10.1093/aje/kwab087}. It also includes general tools for evaluating differences in adjacency matrices, which can be used for evaluating performance of causal discovery procedures.
}
\section{System requirements}{

Requires Java (version 21 or newer) to run the bundled Tetrad Java library.
Requires Cargo (Rust's package manager), rustc >= 1.80.0, xz
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/annennenne/causalDisco}
  \item Report bugs at \url{https://github.com/annennenne/causalDisco/issues}
}

}
\author{
\strong{Maintainer}: Anne Helby Petersen \email{ahpe@sund.ku.dk}

}
