% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDisco_search.R
\name{causalDiscoSearch}
\alias{causalDiscoSearch}
\title{Class for causalDisco search algorithms}
\description{
This class implements the search algorithms from the causalDisco package,
which wraps and adds temporal order to pcalg algorithms.
It allows to set the data, sufficient statistics, test, score, and algorithm.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-causalDiscoSearch-new}{\code{causalDiscoSearch$new()}}
\item \href{#method-causalDiscoSearch-set_params}{\code{causalDiscoSearch$set_params()}}
\item \href{#method-causalDiscoSearch-set_data}{\code{causalDiscoSearch$set_data()}}
\item \href{#method-causalDiscoSearch-set_suff_stat}{\code{causalDiscoSearch$set_suff_stat()}}
\item \href{#method-causalDiscoSearch-set_test}{\code{causalDiscoSearch$set_test()}}
\item \href{#method-causalDiscoSearch-set_score}{\code{causalDiscoSearch$set_score()}}
\item \href{#method-causalDiscoSearch-set_alg}{\code{causalDiscoSearch$set_alg()}}
\item \href{#method-causalDiscoSearch-set_knowledge}{\code{causalDiscoSearch$set_knowledge()}}
\item \href{#method-causalDiscoSearch-run_search}{\code{causalDiscoSearch$run_search()}}
\item \href{#method-causalDiscoSearch-clone}{\code{causalDiscoSearch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-causalDiscoSearch-new"></a>}}
\if{latex}{\out{\hypertarget{method-causalDiscoSearch-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{causalDiscoSearch$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-causalDiscoSearch-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-causalDiscoSearch-set_params}{}}}
\subsection{Method \code{set_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{causalDiscoSearch$set_params(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-causalDiscoSearch-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-causalDiscoSearch-set_data}{}}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{causalDiscoSearch$set_data(data, set_suff_stat = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-causalDiscoSearch-set_suff_stat"></a>}}
\if{latex}{\out{\hypertarget{method-causalDiscoSearch-set_suff_stat}{}}}
\subsection{Method \code{set_suff_stat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{causalDiscoSearch$set_suff_stat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-causalDiscoSearch-set_test"></a>}}
\if{latex}{\out{\hypertarget{method-causalDiscoSearch-set_test}{}}}
\subsection{Method \code{set_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{causalDiscoSearch$set_test(method, alpha = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-causalDiscoSearch-set_score"></a>}}
\if{latex}{\out{\hypertarget{method-causalDiscoSearch-set_score}{}}}
\subsection{Method \code{set_score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{causalDiscoSearch$set_score(method, params = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-causalDiscoSearch-set_alg"></a>}}
\if{latex}{\out{\hypertarget{method-causalDiscoSearch-set_alg}{}}}
\subsection{Method \code{set_alg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{causalDiscoSearch$set_alg(method)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-causalDiscoSearch-set_knowledge"></a>}}
\if{latex}{\out{\hypertarget{method-causalDiscoSearch-set_knowledge}{}}}
\subsection{Method \code{set_knowledge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{causalDiscoSearch$set_knowledge(knowledge_obj, directed_as_undirected = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-causalDiscoSearch-run_search"></a>}}
\if{latex}{\out{\hypertarget{method-causalDiscoSearch-run_search}{}}}
\subsection{Method \code{run_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{causalDiscoSearch$run_search(data = NULL, set_suff_stat = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-causalDiscoSearch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-causalDiscoSearch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{causalDiscoSearch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
