% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare two tpdag or tskeleton objects}
\usage{
compare(x, y = NULL)
}
\arguments{
\item{x}{First object}

\item{y}{Second object (optional)}
}
\value{
A list with entries: \code{$nedges1} (the number of
edges in the first object), \code{$nedges2} (the number of edges
in the second object), \code{$psi1} (the test significance level
of the first object), \code{$psi2} (the test significance level of
the second object), \code{$nadded} (the number of additional edges in
object 2, relative to object 1), and \code{nremoved} (the number of
absent edges in object 2, relative to object 1).
}
\description{
Compare edges in two tpdag objects or two tskeleton objects.
Note that they should be based on the same variables.
Only edge absence/presence is compared, not edge orientation.
}
