% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional-independence.R
\name{cor_test}
\alias{cor_test}
\title{Test for Vanishing Partial Correlations}
\usage{
cor_test(x, y, conditioning_set, suff_stat)
}
\arguments{
\item{x}{Index of x variable.}

\item{y}{Index of y variable.}

\item{conditioning_set}{Index vector of conditioning variable(s), possibly \code{NULL}.}

\item{suff_stat}{Sufficient statistic; list with data,
binary variables and order.}
}
\value{
A numeric, which is the p-value of the test.
}
\description{
This function simply calls the \code{\link[pcalg:condIndFisherZ]{pcalg::gaussCItest()}} function from the \pkg{pcalg} package.
}
