% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corTest.R
\name{corTest}
\alias{corTest}
\title{Test for vanishing partial correlations}
\usage{
corTest(x, y, S, suffStat)
}
\arguments{
\item{x}{Index of x variable}

\item{y}{Index of y variable}

\item{S}{Index of S variable(s), possibly NULL}

\item{suffStat}{Sufficient statistic; list with data,
binary variables and order.}
}
\value{
A numeric, which is the p-value of the test.
}
\description{
This function simply calls the \code{\link[pcalg]{gaussCItest}}
function from the \code{pcalg} package.
}
