% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{deparse_knowledge}
\alias{deparse_knowledge}
\title{Deparse a knowledge object to knowledge() mini-DSL code}
\usage{
deparse_knowledge(.kn, df_name = NULL)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{df_name}{Optional name of the data frame you used
(used as the first argument to \code{knowledge()}).  If \code{NULL},
\code{knowledge()} is called with no data frame.}
}
\value{
A single string (with newlines) of R code.
}
\description{
Given a \code{knowledge} object, return a single string containing
the R code (using \code{knowledge()}, \code{tier()}, \code{forbidden()}, \code{required()})
that would rebuild that same object.
}
