% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco.R
\name{disco}
\alias{disco}
\title{Disco!!}
\usage{
disco(data, method, knowledge = NULL)
}
\arguments{
\item{data}{A data frame}

\item{method}{A \code{disco_method} object representing a causal discovery
algorithm.}

\item{knowledge}{A \code{knowledge} object to be incorporated into the disco method. If \code{NULL} (default), the method is
applied without additional knowledge.}
}
\value{
A \code{caugi} and a \code{knowledge} (\code{knowledgeable_caugi}) object.
}
\description{
Run a causal discovery method on a data frame.
}
\examples{
### disco() example ###
\dontrun{
data("tpcExample")

# define background knowledge
kn <- knowledge(
  tpcExample,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

# use PC (for example)
my_pc <- pc(engine = "tetrad", test = "fisher_z", alpha = 0.01)

disco(data = tpcExample, method = my_pc)
}
}
