% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco.R
\name{disco}
\alias{disco}
\title{Disco!!}
\usage{
disco(data, method, knowledge = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{method}{A \code{disco_method} object representing a causal discovery
algorithm. Available methods are
\itemize{
\item \code{\link[=pc]{pc()}} - PC algorithm,
\item \code{\link[=fci]{fci()}} - FCI algorithm,
\item \code{\link[=ges]{ges()}} - GES algorithm,
\item \code{\link[=tges]{tges()}} - TGES algorithm,
\item \code{\link[=tpc]{tpc()}} - TPC algorithm.
}}

\item{knowledge}{A \code{knowledge} object to be incorporated into the disco method. If \code{NULL} (default), the method is
applied without additional knowledge.}
}
\value{
A \code{caugi} and a \code{knowledge} (\code{knowledgeable_caugi}) object.
}
\description{
Run a causal discovery method on a data frame.
}
\details{
For specific details on the supported algorithms, scores, tests, and parameters for each engine, see:
\itemize{
\item \code{\link{TetradSearch}} for \pkg{Tetrad},
\item \code{\link{PcalgSearch}} for \pkg{pcalg},
\item \code{\link{BnlearnSearch}} for \pkg{bnlearn},
\item \code{\link{CausalDiscoSearch}} for \pkg{causalDisco}.
}
}
\examples{
data("tpc_example")

# use pc with engine bnlearn and test fisher_z
my_pc <- pc(engine = "bnlearn", test = "fisher_z", alpha = 0.01)
pc_bnlearn <- disco(data = tpc_example, method = my_pc)
plot(pc_bnlearn)

# define tiered background knowledge
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

# use gs with engine bnlearn and test cor and tiered background knowledge
my_pc_tiered <- pc(engine = "bnlearn", test = "cor", alpha = 0.01)
pc_tiered_bnlearn <- disco(
  data = tpc_example,
  method = my_pc_tiered,
  knowledge = kn
)
plot(pc_tiered_bnlearn)
}
