% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco_method.R
\name{disco_method}
\alias{disco_method}
\title{Wrap a runner list into a “disco_method” closure}
\usage{
disco_method(builder, method_class)
}
\description{
Wrap a runner list into a “disco_method” closure
}
\keyword{internal}
