% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.amat.cpdag}
\alias{discography.amat.cpdag}
\title{Convert amat.cpdag object to a discography object.}
\usage{
\method{discography}{amat.cpdag}(x, nodes = NULL, ...)
}
\description{
Convert amat.cpdag object to a discography object.
}
