% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.amat.pag}
\alias{discography.amat.pag}
\title{Convert amat.pag object to a discography object.}
\usage{
\method{discography}{amat.pag}(x, nodes = NULL, ...)
}
\description{
Convert amat.pag object to a discography object.
}
