% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.amat}
\alias{discography.amat}
\title{Convert amat object to a discography object.}
\usage{
\method{discography}{amat}(x, nodes = NULL, ...)
}
\description{
Convert amat object to a discography object.
}
