% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.bn}
\alias{discography.bn}
\title{Convert bnlearn object to a discography object.}
\usage{
\method{discography}{bn}(x, nodes = names(x$nodes), ...)
}
\description{
Convert bnlearn object to a discography object.
}
