% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discography.R
\name{discography.graphNEL}
\alias{discography.graphNEL}
\title{Convert graphNEL object to a discography object.}
\usage{
\method{discography}{graphNEL}(x, nodes = graph::nodes(x), ...)
}
\arguments{
\item{x}{A causal-graph object (see methods below).}

\item{nodes}{Optional character vector giving the desired node order.
If \code{NULL}, the unique node names are sorted alphabetically.}

\item{...}{Passed on to methods.}
}
\value{
A \code{"discography"} tibble.
}
\description{
Convert graphNEL object to a discography object.
}
