% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-causaldisco-algs.R
\name{distribute_engine_args}
\alias{distribute_engine_args}
\title{Validate and distribute engine arguments}
\usage{
distribute_engine_args(search, args, engine, alg)
}
\arguments{
\item{search}{R6 object, either \code{TetradSearch}, \code{BnlearnSearch}, \code{PcalgSearch}, or \code{CausalDiscoSearch}.}

\item{args}{List of arguments to distribute}

\item{engine}{Engine identifier, either "tetrad", "bnlearn", "pcalg", or "causalDisco"}

\item{alg}{Algorithm name}
}
\description{
Validate and distribute engine arguments
}
\seealso{
Other Extending causalDisco: 
\code{\link{new_disco_method}()},
\code{\link{register_tetrad_algorithm}()},
\code{\link{reset_tetrad_alg_registry}()}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
