% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{.add_edges}
\alias{.add_edges}
\title{Add one or many edges to a knowledge object}
\usage{
.add_edges(.kn, status, from, to)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{status}{A string, either "forbidden" or "required".}

\item{from}{A tidyselect specification or character vector of variable names.}

\item{to}{A tidyselect specification or character vector of variable names.}
}
\description{
Add one or many edges to a knowledge object
}
\keyword{internal}
