% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{.add_edges}
\alias{.add_edges}
\title{Add one or many edges to a knowledge object}
\usage{
.add_edges(.kn, status, from, to, remove_self_loops = TRUE)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{status}{A string, either "forbidden" or "required".}

\item{from}{A tidyselect specification or character vector of variable names.}

\item{to}{A tidyselect specification or character vector of variable names.}

\item{remove_self_loops}{Logical. If \code{TRUE}, self-loops are removed.
Defaults to \code{FALSE}.}
}
\value{
The updated \code{knowledge} object.
}
\description{
Add one or many edges to a knowledge object
}
\examples{
# add a block of forbidden edges using internal helper
data(tpcExample)

kn <- knowledge(tpcExample)

# add forbidden edges between youth_* and child_*
kn <- causalDisco:::.add_edges(kn, "forbidden",
  from = c("youth_x3", "youth_x4"),
  to = c("child_x1", "child_x2")
)

print(kn)
}
\keyword{internal}
