% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_testing.R
\docType{data}
\name{.allowed_edge_types}
\alias{.allowed_edge_types}
\title{Supported edge-type strings}
\format{
An object of class \code{character} of length 6.
}
\usage{
.allowed_edge_types
}
\description{
Supported edge-type strings
}
\keyword{internal}
