% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{.edge_verb}
\alias{.edge_verb}
\title{Handle \code{forbid_edge()} and \code{require_edge()} verbs}
\usage{
.edge_verb(.kn, status, from_quo, to_quo = NULL)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{status}{A string, either "forbidden" or "required".}

\item{from}{A tidyselect specification or a variable name string or symbol.}

\item{to}{A tidyselect specification or a variable name string or symbol.}
}
\description{
Help \code{forbid_edge()} and \code{require_edge()} to handle both
tidyselect and formula inputs as well as standard \code{to}, \code{from} arguments.
Passes arguments correctly to \code{.add_edges()}.
}
\keyword{internal}
