% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{.formula_vars}
\alias{.formula_vars}
\title{Extract variable names from the RHS of a \code{tier()} formula}
\usage{
.formula_vars(.kn, rhs)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{rhs}{A formula (e.g. \code{1 ~ V1 + V2}).}
}
\description{
Extract variable names from the RHS of a \code{tier()} formula
}
\keyword{internal}
