% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_testing.R
\name{.formula_vars}
\alias{.formula_vars}
\title{Extract variable names from the RHS of a \code{tier()} formula}
\usage{
.formula_vars(rhs, .kn)
}
\description{
Extract variable names from the RHS of a \code{tier()} formula
}
\keyword{internal}
