% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{.new_knowledge}
\alias{.new_knowledge}
\title{Create a \code{knowledge} object}
\usage{
.new_knowledge(vars = character(), frozen = FALSE)
}
\arguments{
\item{vars}{Character vector of variable names.  Defaults to empty.}

\item{frozen}{Logical. If \code{TRUE}, no new variables can be added. Defaults to \code{FALSE}.}
}
\value{
An S3 object of class \code{"knowledge"}.
}
\description{
Create a \code{knowledge} object
}
\examples{
### .new_knowledge() example ###

# create a bare knowledge object using .new_knowledge()
kn1 <- try(causalDisco:::.new_knowledge(vars = c("A", "B", "C"), frozen = TRUE))

# create the knowledge object in the intended way
df <- data.frame(A = rnorm(10), B = rnorm(10), C = rnorm(10))
kn2 <- knowledge(df)

identical(kn1, kn2) # TRUE
}
\keyword{internal}
