% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_testing.R
\name{.new_knowledge}
\alias{.new_knowledge}
\title{Create a \code{knowledge} object}
\usage{
.new_knowledge(vars = character())
}
\arguments{
\item{vars}{Character vector of variable names.  Defaults to empty.}
}
\value{
An S3 object of class \code{"knowledge"}.
}
\description{
Create a \code{knowledge} object
}
\keyword{internal}
