% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_testing.R
\name{.validate_forbidden_required}
\alias{.validate_forbidden_required}
\title{Validate that an edge is not simultaneously forbidden \emph{and} required}
\usage{
.validate_forbidden_required(edges_df)
}
\description{
Validate that an edge is not simultaneously forbidden \emph{and} required
}
\keyword{internal}
