% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{.validate_forbidden_required}
\alias{.validate_forbidden_required}
\title{Validate that an edge is not simultaneously forbidden \emph{and} required}
\usage{
.validate_forbidden_required(edges_df)
}
\arguments{
\item{edges_df}{A data frame with columns \code{status}, \code{edge_type}, \code{from},
\code{to}, \code{tier_from}, and \code{tier_to}.}
}
\description{
Validate that an edge is not simultaneously forbidden \emph{and} required
}
\keyword{internal}
