% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeable-caugi.R
\name{edges}
\alias{edges}
\title{Retrieve edges from a \code{knowledgeable_caugi} object}
\usage{
edges(kcg)
}
\arguments{
\item{kcg}{A \code{knowledgeable_caugi} object.}
}
\value{
A tibble containing the edges.
}
\description{
Retrieve edges from a \code{knowledgeable_caugi} object
}
\examples{
data(tpc_example)
cd_tges <- tges(engine = "causalDisco", score = "tbic")
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)
disco_cd_tges <- disco(data = tpc_example, method = cd_tges, knowledge = kn)
edges(disco_cd_tges)
}
