% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges}
\alias{edges}
\title{List of edges in adjacency matrix}
\usage{
edges(amat)
}
\arguments{
\item{amat}{An adjacency matrix.}
}
\description{
Produces a list of edges from an adjacency matrix. The output list consists
of two lists itself, one for oriented edges (\code{$dir}), and one for
unoriented edges (\code{$undir}).
}
