% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-registry.R
\docType{data}
\name{engine_registry}
\alias{engine_registry}
\title{Supported Engines}
\format{
An object of class \code{list} of length 4.
}
\usage{
engine_registry
}
\description{
Available engines: bnlearn, causalDisco, pcalg, tetrad.
}
\keyword{datasets}
