% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{false_omission_rate}
\alias{false_omission_rate}
\title{False Omission Rate}
\usage{
false_omission_rate(truth, guess, type = c("adj", "dir"))
}
\arguments{
\item{truth}{A \code{caugi} object representing the true graph.}

\item{guess}{A \code{caugi} object representing the estimated graph.}

\item{type}{Character string specifying the comparison type:
\itemize{
\item \code{"adj"}: adjacency comparison.
\item \code{"dir"}: orientation comparison conditional on shared adjacencies.
}}
}
\value{
A numeric in [0,1].
}
\description{
Computes false omission rate from two \code{caugi} objects.
It converts the \code{caugi} objects to adjacency matrices and computes
false omission rate as \code{FN/(FN + TN)}, where \code{FN} are false negatives and
\code{TN} are true negatives. If \verb{FN + TN = 0, 1} is returned.
}
\examples{
cg1 <- caugi::caugi(A \%-->\% B + C)
cg2 <- caugi::caugi(B \%-->\% A + C)
false_omission_rate(cg1, cg2, type = "adj")
false_omission_rate(cg1, cg2, type = "dir")

}
\seealso{
Other metrics: 
\code{\link{confusion}()},
\code{\link{f1_score}()},
\code{\link{fdr}()},
\code{\link{g1_score}()},
\code{\link{npv}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{shd}()},
\code{\link{specificity}()}
}
\concept{metrics}
