% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fci.R
\name{fci}
\alias{fci}
\title{Perform causal discovery using the FCI algorithm}
\usage{
fci(engine = "tetrad", test, alpha = 0.05, ...)
}
\arguments{
\item{test}{A procedure for testing conditional independence.
The default, \code{regTest} uses a regression-based information
loss test. Another available option is \code{corTest} which
tests for vanishing partial correlations. User supplied functions
may also be used, see details below about the required syntax.}

\item{...}{Further optional arguments which are passed to
\code{\link[pcalg]{skeleton}} for the skeleton constructing phase.}

\item{output}{One of \code{"pag"} or \code{"fciAlgo"}. If \code{"pag"}
a partial ancestral graph (PAG) object is outputted.
If \code{"fciAlgo"} the PAG is outputted as the
object class \code{\link[pcalg]{fciAlgo-class}} from the pcalg package. This is
intended for compatability with tools from that package.}
}
\description{
This is a wrapper function for the \code{\link[pcalg]{fci}} function as
implemented in the pcalg package. All computations are carried out by the
pcalg package. The default output object however matches that of \link{tfci}, see
this function for details about how the adjacency matrix is encoded.
}
\examples{
# simulate linear Gaussian data w unobserved variable L1
n <- 100
L1 <- rnorm(n) 
X1 <- rnorm(n)
X2 <- L1 + X1 + rnorm(n)
X3 <- X1 + rnorm(n)
X4 <- X3 + L1 + rnorm(n)
d <- data.frame(p1_X1 = X1,
                p1_X2 = X2,
                p2_X3 = X3,
                p2_X4 = X4)

# use FCI algorithm to recover PAG                
fci(d, test = corTest)


}
