% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpc_run.R
\name{find_adjacencies}
\alias{find_adjacencies}
\title{Find adjacencies of a node in an adjacency matrix}
\usage{
find_adjacencies(amatrix, index)
}
\arguments{
\item{amatrix}{Square adjacency matrix.}

\item{index}{Integer index of the node.}
}
\value{
Integer vector of adjacent node indices.
}
\description{
Find adjacencies of a node in an adjacency matrix
}
\keyword{internal}
