% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge_testing.R
\name{forbid_edge}
\alias{forbid_edge}
\alias{require_edge}
\title{Register forbidden / required edges}
\usage{
forbid_edge(.kn, from, to, edge_type = "directed")

require_edge(.kn, from, to, edge_type = "directed")
}
\description{
Register forbidden / required edges
}
