% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{forbid_edge}
\alias{forbid_edge}
\title{Add forbidden edges}
\usage{
forbid_edge(.kn, ...)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{...}{One or more two-sided formulas.}
}
\value{
The updated \code{knowledge} object.
}
\description{
Forbid one or more directed edges.
Each argument \strong{must} be a two–sided formula, e.g. \code{X ~ Y}.
Formulas can use tidy-select on either side, so
\code{forbid_edge(kn, starts_with("X") ~ Y)} forbids every \verb{X_i --> Y}.
}
