% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{forbid_tier_violations}
\alias{forbid_tier_violations}
\title{Forbid all tier violations}
\usage{
forbid_tier_violations(.kn)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}
}
\value{
The same \code{knowledge} object with new forbidden edges added.
}
\description{
Given a \code{knowledge} object with variables already assigned to tiers,
forbids every directed edge that runs from a higher-numbered tier down
into a lower-numbered tier.
}
\examples{
### forbid_tier_violations() example ###

# automatically forbid edges that go from later tiers to earlier tiers
data(tpcExample)

kn <- knowledge(
  head(tpcExample),
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

# turn all tier violations to forbidden edges
kn2 <- forbid_tier_violations(kn)

print(kn)
print(kn2)
}
