% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{forbid_tier_violations}
\alias{forbid_tier_violations}
\title{Forbid all tier violations}
\usage{
forbid_tier_violations(.kn)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}
}
\value{
The same \code{knowledge} object with new forbidden edges added.
}
\description{
Given a \code{knowledge} object with variables already assigned to tiers,
forbids every directed edge that runs from a higher-numbered tier down
into a lower-numbered tier.
}
\examples{
### forbid_tier_violations() example ###

# automatically forbid edges that go from later tiers to earlier tiers
data(tpc_example)

kn <- knowledge(
  head(tpc_example),
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

# turn all tier violations to forbidden edges
kn2 <- forbid_tier_violations(kn)

print(kn)
print(kn2)
}
\seealso{
Other knowledge functions: 
\code{\link{+.knowledge}()},
\code{\link{add_exogenous}()},
\code{\link{add_tier}()},
\code{\link{add_to_tier}()},
\code{\link{add_vars}()},
\code{\link{as_bnlearn_knowledge}()},
\code{\link{as_pcalg_constraints}()},
\code{\link{as_tetrad_knowledge}()},
\code{\link{deparse_knowledge}()},
\code{\link{forbid_edge}()},
\code{\link{get_tiers}()},
\code{\link{knowledge}()},
\code{\link{remove_edges}()},
\code{\link{remove_tiers}()},
\code{\link{remove_vars}()},
\code{\link{reorder_tiers}()},
\code{\link{reposition_tier}()},
\code{\link{require_edge}()},
\code{\link{seq_tiers}()},
\code{\link{unfreeze}()}
}
\concept{knowledge}
\concept{knowledge functions}
