% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\docType{class}
\name{GaussL0penIntScoreORDER-class}
\alias{GaussL0penIntScoreORDER-class}
\title{Constructor
Calculates the local score of a vertex and its parents}
\description{
Constructor
Calculates the local score of a vertex and its parents
}

