% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ges.R
\name{ges}
\alias{ges}
\title{The GES algorithm for causal discovery}
\usage{
ges(engine = c("tetrad", "pcalg"), score, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"tetrad"}}{Tetrad Java library.}
\item{\code{"pcalg"}}{\pkg{pcalg} R package.}
}}

\item{score}{Character; name of the scoring function to use.}

\item{...}{Additional arguments passed to the chosen engine (e.g. test or
algorithm parameters).}
}
\value{
A function of class \code{"ges"} that takes a single argument \code{data}
(a data frame) and returns an \code{caugi} and a \code{knowledge}
(\code{knowledgeable_caugi}) object.
}
\description{
Run the GES algorithm for causal discovery using one of several engines.
}
\examples{
### ges() example ###

data("tpcExample")

kn <- knowledge(
  tpcExample,
  tier(
    child ~ tidyselect::starts_with("child"),
    youth ~ tidyselect::starts_with("youth"),
    oldage ~ tidyselect::starts_with("oldage")
  )
)

# Recommended path using disco()
my_ges <- ges(engine = "tetrad", score = "sem_bic")

disco(tpcExample, my_ges, knowledge = kn)

# or using my_ges directly
my_ges <- my_ges |> set_knowledge(kn)
my_ges(tpcExample)
}
