% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ges.R
\name{ges}
\alias{ges}
\title{GES Algorithm for Causal Discovery}
\usage{
ges(engine = c("tetrad", "pcalg"), score, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"tetrad"}}{\pkg{Tetrad} Java library.}
\item{\code{"pcalg"}}{\pkg{pcalg} R package.}
}}

\item{score}{Character; name of the scoring function to use.}

\item{...}{Additional arguments passed to the chosen engine (e.g. score and
algorithm parameters).}
}
\value{
A function of class \code{"ges"} that takes a single argument \code{data}
(a data frame) and returns a \code{caugi} (of class "PDAG") and a \code{knowledge}
(\code{knowledgeable_caugi}) object.
}
\description{
Run the GES algorithm for causal discovery using one of several engines.
}
\details{
For specific details on the supported scores, and parameters for each engine, see:
\itemize{
\item \code{\link{TetradSearch}} for \pkg{Tetrad} (note, Tetrad refers to it as "fges"),
\item \code{\link{PcalgSearch}} for \pkg{pcalg}.
}
}
\examples{
data(tpc_example)

#### Using pcalg engine ####
# Recommended path using disco()
ges_pcalg <- ges(engine = "pcalg", score = "sem_bic")
disco(tpc_example, ges_pcalg)

# or using ges_pcalg directly
ges_pcalg(tpc_example)

# With all algorithm arguments specified
ges_pcalg <- ges(
  engine = "pcalg",
  score = "sem_bic",
  adaptive = "vstructures",
  phase = "forward",
  iterate = FALSE,
  maxDegree = 3,
  verbose = FALSE
)
disco(tpc_example, ges_pcalg)


#### Using tetrad engine with tier knowledge ####
# Requires Tetrad to be installed
if (check_tetrad_install()$installed && check_tetrad_install()$java_ok) {
  kn <- knowledge(
    tpc_example,
    tier(
      child ~ tidyselect::starts_with("child"),
      youth ~ tidyselect::starts_with("youth"),
      oldage ~ tidyselect::starts_with("oldage")
    )
  )

  # Recommended path using disco()
  ges_tetrad <- ges(engine = "tetrad", score = "sem_bic")
  disco(tpc_example, ges_tetrad, knowledge = kn)

  # or using ges_tetrad directly
  ges_tetrad <- ges_tetrad |> set_knowledge(kn)
  ges_tetrad(tpc_example)
}

# With all algorithm arguments specified
if (check_tetrad_install()$installed && check_tetrad_install()$java_ok) {
  ges_tetrad <- ges(
    engine = "tetrad",
    score = "ebic",
    symmetric_first_step = TRUE,
    max_degree = 3,
    parallelized = TRUE,
    faithfulness_assumed = TRUE
  )
  disco(tpc_example, ges_tetrad)
}
}
