% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{graph2amat}
\alias{graph2amat}
\title{Convert graphNEL object to adjacency matrix}
\usage{
graph2amat(graph)
}
\arguments{
\item{graph}{A graphNEL object.}
}
\description{
Convert graphNEL object to adjacency matrix
}
