% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasp.R
\name{grasp}
\alias{grasp}
\title{GRaSP Algorithm for Causal Discovery}
\usage{
grasp(engine = "tetrad", score, test, alpha = 0.05, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"tetrad"}}{\pkg{Tetrad} Java library.}
}}

\item{score}{Character; name of the scoring function to use.}

\item{test}{Character; name of the conditional‐independence test.}

\item{alpha}{Numeric; significance level for the CI tests.}

\item{...}{Additional arguments passed to the chosen engine (e.g. score and
algorithm parameters).}
}
\value{
A function of class \code{"grasp"} that takes a single argument \code{data}
(a data frame) and returns a \code{caugi} (of class "PDAG") and a \code{knowledge}
(\code{knowledgeable_caugi}) object.
}
\description{
Run the GRaSP algorithm for causal discovery using one of several engines.
}
\details{
For specific details on the supported scores, and parameters for each engine, see:
\itemize{
\item \code{\link{TetradSearch}} for \pkg{Tetrad}.
}
}
\examples{
data(tpc_example)

# Requires Tetrad to be installed
if (check_tetrad_install()$installed && check_tetrad_install()$java_ok) {
  # Recommended path using disco()
  grasp_tetrad <- grasp(
    engine = "tetrad",
    test = "fisher_z",
    score = "sem_bic",
    alpha = 0.05
  )
  disco(tpc_example, grasp_tetrad)

  # or using grasp_tetrad directly
  grasp_tetrad(tpc_example)
}

#### With tier knowledge ####
if (check_tetrad_install()$installed && check_tetrad_install()$java_ok) {
  kn <- knowledge(
    tpc_example,
    tier(
      child ~ tidyselect::starts_with("child"),
      youth ~ tidyselect::starts_with("youth"),
      oldage ~ tidyselect::starts_with("oldage")
    )
  )

  # Recommended path using disco()
  grasp_tetrad <- grasp(
    engine = "tetrad",
    test = "fisher_z",
    score = "sem_bic",
    alpha = 0.05
  )
  disco(tpc_example, grasp_tetrad, knowledge = kn)

  # or using grasp_tetrad directly
  grasp_tetrad <- grasp_tetrad |> set_knowledge(kn)
  grasp_tetrad(tpc_example)
}
}
