% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-iamb-family.R, R/iamb-family.R
\name{iamb-family}
\alias{iamb-family}
\alias{iamb}
\alias{iamb_fdr}
\alias{fast_iamb}
\alias{inter_iamb}
\title{IAMB Family of Causal Discovery Algorithms}
\usage{
iamb(engine = c("bnlearn"), test, alpha = 0.05, ...)

iamb_fdr(engine = c("bnlearn"), test, alpha = 0.05, ...)

fast_iamb(engine = c("bnlearn"), test, alpha = 0.05, ...)

inter_iamb(engine = c("bnlearn"), test, alpha = 0.05, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"bnlearn"}}{\pkg{bnlearn} R package.}
}}

\item{test}{Character; name of the conditional‐independence test.}

\item{alpha}{Numeric; significance level for the CI tests.}

\item{...}{Additional arguments passed to the chosen engine (e.g., test or algorithm parameters).}
}
\value{
Each function returns a \code{caugi} object (of class "PDAG") and a \code{knowledge} object.
}
\description{
Functions for causal discovery using variants of the Incremental Association algorithm:
\itemize{
\item \code{iamb}: Incremental Association (IAMB)
\item \code{inter_iamb}: Interleaved Incremental Association (Inter-IAMB)
\item \code{iamb_fdr}: Incremental Association with FDR (IAMB-FDR)
\item \code{fast_iamb}: Fast Incremental Association (Fast-IAMB)
}
}
\details{
Each function supports the same engines and parameters. For details on tests
and parameters for each engine, see:
\itemize{
\item \code{\link{BnlearnSearch}} for \pkg{bnlearn}.
}
}
\examples{
data(tpc_example)

kn <- knowledge(
  tpc_example,
  starts_with("child") \%-->\% starts_with("youth")
)

##### iamb #####

# Recommended path using disco()
iamb_bnlearn <- iamb(engine = "bnlearn", test = "fisher_z", alpha = 0.05)
disco(tpc_example, iamb_bnlearn, knowledge = kn)

# or using iamb_bnlearn directly
iamb_bnlearn <- iamb_bnlearn |> set_knowledge(kn)
iamb_bnlearn(tpc_example)


# With all algorithm arguments specified
iamb_bnlearn <- iamb(
  engine = "bnlearn",
  test = "fisher_z",
  alpha = 0.05,
  max.sx = 2,
  debug = FALSE,
  undirected = TRUE
)

disco(tpc_example, iamb_bnlearn)

##### iamb_fdr #####

iamb_fdr_bnlearn <- iamb_fdr(
  engine = "bnlearn",
  test = "fisher_z",
  alpha = 0.05
)
disco(tpc_example, iamb_fdr_bnlearn, knowledge = kn)

##### fast_iamb #####

fast_iamb_bnlearn <- fast_iamb(
  engine = "bnlearn",
  test = "fisher_z",
  alpha = 0.05
)
disco(tpc_example, fast_iamb_bnlearn, knowledge = kn)

#### inter_iamb #####

inter_iamb_bnlearn <- inter_iamb(
  engine = "bnlearn",
  test = "fisher_z",
  alpha = 0.05
)
disco(tpc_example, inter_iamb_bnlearn, knowledge = kn)
}
\seealso{
Other causal discovery algorithms: 
\code{\link{boss}()},
\code{\link{boss_fci}()},
\code{\link{fci}()},
\code{\link{ges}()},
\code{\link{gfci}()},
\code{\link{grasp}()},
\code{\link{grasp_fci}()},
\code{\link{gs}()},
\code{\link{hc}()},
\code{\link{pc}()},
\code{\link{sp_fci}()},
\code{\link{tabu}()},
\code{\link{tfci}()},
\code{\link{tges}()},
\code{\link{tpc}()}
}
\concept{causal discovery algorithms}
\concept{cd_algorithms}
