% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_java.R
\name{install_java}
\alias{install_java}
\title{Install Temurin JDK 25}
\usage{
install_java(force = FALSE)
}
\arguments{
\item{force}{Logical; if TRUE, forces re-installation even if JDK is already installed.}
}
\description{
This function installs the Eclipse Temurin JDK 25 on the user's system. It works
on Windows and macOS operating systems. For Linux we assume the user can install
Java via their package manager.
The JDK is installed in the user's home directory.
}
\examples{
\dontrun{
install_java()
}
}
