% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tetrad.R
\name{install_tetrad}
\alias{install_tetrad}
\title{Install Tetrad GUI}
\usage{
install_tetrad(
  version = getOption("causalDisco.tetrad.version"),
  dir = NULL,
  set_renviron = TRUE,
  force = FALSE
)
}
\arguments{
\item{version}{Character. The version of Tetrad to install.
Default is "7.6.10".
Use \code{options(causalDisco.tetrad.version = "...")} to change the default version.}

\item{dir}{Character. The directory to install Tetrad into. Default is "~/tetrad".}

\item{set_renviron}{Logical. Whether to set the TETRAD_DIR in .Renviron. Default is TRUE.}

\item{force}{Logical. Whether to force re-download if the file already exists. Default is FALSE.}
}
\value{
Invisible character string of the path to the downloaded JAR file.
}
\description{
This function downloads and installs the Tetrad GUI JAR file to a specified directory.
It also sets the TETRAD_DIR environment variable for future R sessions.
}
\examples{
\dontrun{
install_tetrad()
}
}
