% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{is_knowledge}
\alias{is_knowledge}
\title{Verify that an object is a knowledge}
\usage{
is_knowledge(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check that the object is a \code{knowledge} object. Mostly
for internal use in causalDisco.
}
\keyword{internal}
