% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{knowledge}
\alias{knowledge}
\title{Creates a Knowledge Object by calling the function below.
Example:
my_knowledge <- knowledge(
tier(
1, c("V1", "V2", "V3"),
2, c("V4", "V5", "V6")
),
forbidden("V1", "V6"), # forbidden and required can be called like this
forbidden(c("V2", "V6")), # or like this
required(c("V1", "V2"),
c("V2", "V3")) # or like this
)}
\usage{
knowledge(...)
}
\description{
Creates a Knowledge Object by calling the function below.
Example:
my_knowledge <- knowledge(
tier(
1, c("V1", "V2", "V3"),
2, c("V4", "V5", "V6")
),
forbidden("V1", "V6"), # forbidden and required can be called like this
forbidden(c("V2", "V6")), # or like this
required(c("V1", "V2"),
c("V2", "V3")) # or like this
)
}
