% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{knowledge}
\alias{knowledge}
\title{Knowledge Mini-DSL constructor (\code{tier()}, \code{forbidden()}, \code{required()})}
\usage{
knowledge(...)
}
\value{
A populated \code{knowledge} object.
}
\description{
Accepts an optional data frame followed by calls built from \strong{formulas}:
\itemize{
\item \code{tier( 1 ~ V1 + V2, exposure ~ E )}
\item \code{forbidden( V1 ~ V4, V2 ~ V4, edge_type = "undirected" )}
\item \code{required ( V1 ~ V2 )}
}
}
\details{
The first argument can be a data frame, which will be used to populate the
\code{knowledge} object with variable names. If you later add variables with
add_* verbs, this will throw a warning, since the knowledge object will
be \emph{frozen}. You can unfreeze a knowledge object by using the function
\code{unfreeze(knowledge)}.

If no data frame is provided, the
\code{knowledge} object will be empty until variables are added with \code{tier()},
\code{forbidden()}, or \code{required()}. You can also populate the object with
the \code{add_vars()} verb.

\code{tier()} assigns variables to tiers. Tiers are internally numbered, starting
with 1. If you provide a numeric literal, it will be used as the tier index.
If you provide a symbol or string, it will be used as a label. The order of
the provided tiers will be used as the order of the tiers in the object,
unless tiers are specified with numeric literals. This function takes
formulas as input. The left-hand side of the formula is the \emph{tier} and the
right-hand side is the \emph{variables}. You can also use tidyselect syntax to
specify the variables. For example, \code{tier(1 ~ starts_with("V"))} will assign
all variables starting with "V" to tier 1.

The \code{forbidden()} and \code{required()} functions add edges to the knowledge
object. The edges are added as directed edges by default, but you can specify
the \code{edge_type} argument to use undirected or bidirected edges. There is
currently also support for PAG-type edges, but these are generally not
supported in other packages. These functions also take formulas as input.
The left-hand side of the formula is the \emph{from} variable, and the right-hand
side is the \emph{to} variable.
}
