% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi_with_knowledge.R
\name{knowledgeable_caugi}
\alias{knowledgeable_caugi}
\title{A \code{caugi} with an attached \code{knowledge} object}
\usage{
knowledgeable_caugi(graph, kn = knowledge(), class = "PDAG")
}
\arguments{
\item{graph}{A causal graph object}

\item{kn}{A \code{knowledge} object. Default is empty knowledge object.}

\item{class}{A string describing the graph class.}

\item{nodes}{Optional character vector all the node names.}
}
\value{
A \code{caugi} and a \code{knowledge} object in a list.
}
\description{
This S3 class wraps \code{caugi} graph object and a \code{knowledge} object. It is the
output object of causal discovery methods used in \code{causalDisco}.
}
\details{
The conversion from any graph type to a \code{caugi} is handled by the \code{caugi}
package.
}
\seealso{
\code{\link[caugi:caugi]{caugi::caugi()}}
}
