% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketikz.R
\name{make_tikz}
\alias{make_tikz}
\title{Generate Latex tikz code for plotting a temporal DAG, PDAG or PAG.}
\usage{
make_tikz(
  model,
  x_jit = 2,
  y_jit = 2,
  mark_periods = TRUE,
  period_gap = 4,
  annotate_edges = NULL,
  add_axis = TRUE,
  var_labels = NULL,
  period_labels = NULL,
  annotation_labels = NULL,
  clipboard = TRUE,
  raw_out = FALSE,
  color_annotate = NULL,
  bend_edges = FALSE
)
}
\arguments{
\item{model}{\code{tpdag}, \code{tskeleton}, \code{tpag}, or \code{tamat} object to plot.}

\item{x_jit}{How much should nodes within a period be jittered horizontally.}

\item{y_jit}{Vertical distance between nodes within a period.}

\item{mark_periods}{If \code{TRUE}, gray boxes are drawn behind each period.}

\item{period_gap}{Horizontal gap between different periods.}

\item{annotate_edges}{If \code{TRUE}, add a text annotation to edges. If \code{annotation_labels}
are supplied, these labels will be used. Otherwise, the value in the inputted adjacency matrix corresponding
to the edge will be used. Cannot be used for \code{tpag}
input objects (or \code{ag} amat types).}

\item{add_axis}{If \code{TRUE}, a horizontal axis with period labels are added.}

\item{var_labels}{Optional labels for nodes (variables). Should be given as a named list, where
the name is the variable name, and the entry is the label, e.g. \code{list(vname = "Label for vname")}.}

\item{period_labels}{Optional labels for periods. Should be given as a named list, where
the name is the period name (as stored in the \code{tamat}), and the entry is the label,
e.g. \code{list(periodname = "Label for period")}.}

\item{annotation_labels}{Optional labels for edge annotations. Only used if \code{annotate_edges = TRUE}.
Should be given as a named list, where the name is the edge annotation (as stored in the \code{tamat}),
and the entry is the label, e.g. \code{list(h = "High")}.}

\item{clipboard}{If \code{TRUE}, the tikz code is not printed, but instead copied to the clipboard,
so it can easily be pasted into a Latex document.}

\item{raw_out}{If \code{TRUE}, the tikz code is only returned as a character vector.}

\item{color_annotate}{Named list of colors to use to mark edge annotations instead of labels. This
overrules \code{annotate_edges} and both are not available at the same time. The list should be given with
annotations as names and colors as entries, e.g. \code{list(h = "blue")}. Cannot be used for \code{tpag}
input objects (or \code{ag} amat types).}

\item{bend_edges}{If \code{TRUE}, all edges are bend 10 degrees to the right, thereby avoiding having edges exactly
on top of each other.}
}
\value{
Silently returns a character vector with lines of tikz code. The function
furthermore has a side-effect. If \code{clipboard = TRUE}, the side-effect is that the tikz
code is also copied to the clipboard. If \code{clipboard = FALSE}, the tikz code is instead printed
in the console.
}
\description{
Generate Latex tikz code for plotting a temporal DAG, PDAG or PAG.
}
\details{
Note that it is necessary to read in relevant tikz libraries in the
Latex preamble. The relevant lines of code are (depending a bit on parameter settings): \cr
\code{\\usepackage{tikz}} \cr
\code{\\usetikzlibrary{arrows.meta,arrows,shapes,decorations,automata,backgrounds,petri}} \cr
\code{\\usepackage{pgfplots}}
}
\examples{

# Make tikz figure code from tpdag, print code to screen
data(tpc_example)
kn <- knowledge(
  tpc_example,
  tier(
    child ~ tidyselect::starts_with("child"),
    youth ~ tidyselect::starts_with("youth"),
    oldage ~ tidyselect::starts_with("oldage")
  )
)

tpdag_example <- tpc_run(tpc_example,
  kn,
  alpha = 0.01,
  test = cor_test,
  output = "tpdag"
)
make_tikz(tpdag_example, clipboard = FALSE)

# Make tikz figure code from tamat, copy code to clipboard
dag_example <- sim_dag(5)
rownames(dag_example) <- colnames(dag_example) <- c(
  "child_x", "child_y",
  "child_z", "adult_x",
  "adult_y"
)
output_tamat <- tamat(dag_example, order = c("child", "adult"))
\dontrun{
make_tikz(output_tamat)
}

}
