% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{maxnedges}
\alias{maxnedges}
\title{Compute maximal number of edges for graph}
\usage{
maxnedges(p)
}
\arguments{
\item{p}{The number of nodes in the graph}
}
\value{
A numeric.
}
\description{
Computes the number of edges a graph with \code{p} nodes will have if its
fully connected.
}
