% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-causaldisco-algs.R
\name{new_disco_method}
\alias{new_disco_method}
\title{Add a New causalDisco Method}
\usage{
new_disco_method(builder, name, engine, graph_class)
}
\arguments{
\item{builder}{A function returning a runner}

\item{name}{Algorithm name}

\item{engine}{Engine identifier}

\item{graph_class}{Output graph class}
}
\value{
A function of class \code{"disco_method"} that takes a single argument
\code{data} (a data frame) and returns a \code{caugi} object.
}
\description{
This function allows you to create a new causal discovery method that can be used with the \code{\link[=disco]{disco()}} function.
You provide a builder function that constructs a runner object, along with metadata about the algorithm, and it
returns a closure that can be called with a data frame to perform causal discovery and return a \code{caugi} object.
}
\seealso{
Other Extending causalDisco: 
\code{\link{distribute_engine_args}()},
\code{\link{register_tetrad_algorithm}()},
\code{\link{reset_tetrad_alg_registry}()}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
