% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeable-caugi.R
\name{new_knowledgeable_caugi}
\alias{new_knowledgeable_caugi}
\title{Create a new \code{knowledgeable_caugi} object}
\usage{
new_knowledgeable_caugi(cg, kn)
}
\arguments{
\item{cg}{A \code{caugi} object}

\item{kn}{A \code{knowledge} object}
}
\description{
Create a new \code{knowledgeable_caugi} object
}
