% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledgeable_caugi.R
\name{nodes}
\alias{nodes}
\alias{nodes.knowledgeable_caugi}
\title{Retrieve nodes from a \code{knowledgeable_caugi} object}
\usage{
nodes(x)

\method{nodes}{knowledgeable_caugi}(x)
}
\arguments{
\item{x}{A \code{knowledgeable_caugi} object.}
}
\value{
A tibble containing the nodes.
}
\description{
\code{nodes()} extracts the edges from a \code{knowledgeable_caugi} object.
}
\examples{
data("tpc_example")
cd_tges <- tges(engine = "causalDisco", score = "tbic")
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)
disco_cd_tges <- disco(data = tpc_example, method = cd_tges, knowledge = kn)
nodes(disco_cd_tges)
}
