% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caugi-helpers.R
\name{nodes}
\alias{nodes}
\title{Retrieve Nodes}
\usage{
nodes(cg)
}
\arguments{
\item{cg}{A \code{caugi} object.}
}
\value{
A tibble containing the nodes.
}
\description{
This function retrieves the nodes from a \code{caugi} object as a tibble.
}
\examples{
data(tpc_example)
cd_tges <- tges(engine = "causalDisco", score = "tbic")
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)
disco_cd_tges <- disco(data = tpc_example, method = cd_tges, knowledge = kn)
nodes(disco_cd_tges$caugi)
}
