% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{npv}
\alias{npv}
\title{Negative Predictive Value}
\usage{
npv(truth, est, type = c("adj", "dir"))
}
\arguments{
\item{truth}{A \code{caugi} object representing the truth graph.}

\item{est}{A \code{caugi} object representing the estimated graph.}

\item{type}{Character string specifying the comparison type:
\itemize{
\item \code{"adj"}: adjacency comparison.
\item \code{"dir"}: orientation comparison conditional on shared adjacencies.
}}
}
\value{
A numeric in [0,1].
}
\description{
Computes negative predictive value from two PDAG \code{caugi} objects.
It converts the \code{caugi} objects to adjacency matrices and computes
negative predictive value as \code{TN/(TN + FN)}, where \code{TN} are truth negatives and
\code{FN} are false negatives. If \code{TN + FN = 0}, \code{1} is returned.
Only supports \code{caugi} objects with these edge types present \verb{-->}, \verb{<-->}, \verb{---} and no edge.
}
\examples{
cg1 <- caugi::caugi(A \%-->\% B + C)
cg2 <- caugi::caugi(B \%-->\% A + C)
npv(cg1, cg2, type = "adj")
npv(cg1, cg2, type = "dir")

}
\seealso{
Other metrics: 
\code{\link{confusion}()},
\code{\link{evaluate}()},
\code{\link{f1_score}()},
\code{\link{false_omission_rate}()},
\code{\link{fdr}()},
\code{\link{g1_score}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{reexports}},
\code{\link{specificity}()}
}
\concept{metrics}
