% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{NPV}
\alias{NPV}
\title{Negative predictive value}
\usage{
NPV(confusion)
}
\arguments{
\item{confusion}{Confusion matrix as obtained from \link{confusion}}
}
\value{
A numeric in [0,1].
}
\description{
Computes negative predictive value recall from a confusion matrix, see \link{confusion}.
Negative predictive value is defined as TN/(TN + FN), where TN are true negatives and
FN are false negatives. If TP + FN = 0, 0 is returned.
}
