% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.R
\name{pc}
\alias{pc}
\title{The Peter-Clark (PC) algorithm for causal discovery}
\usage{
pc(engine = c("tetrad", "pcalg", "bnlearn"), test, alpha = 0.05, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"tetrad"}}{Tetrad Java library.}
\item{\code{"pcalg"}}{\pkg{pcalg} R package.}
\item{\code{"bnlearn"}}{\pkg{bnlearn} R package.}
}}

\item{test}{Character; name of the conditional‐independence test.}

\item{alpha}{Numeric; significance level for the CI tests.}

\item{...}{Additional arguments passed to the chosen engine (e.g. test or algorithm parameters).}
}
\value{
A function of class \code{"pc"} that takes a single argument \code{data}
(a data frame) and returns an \code{caugi} and a \code{knowledge} object.
}
\description{
Run the PC algorithm for causal discovery using one of several engines.
}
\examples{
### pc() example ###
\dontrun{
data("tpcExample")

kn <- knowledge(
  tpcExample,
  tier(
    child ~ tidyselect::starts_with("child"),
    youth ~ tidyselect::starts_with("youth"),
    oldage ~ tidyselect::starts_with("oldage")
  )
)

# Recommended path using disco()
my_pc <- pc(engine = "tetrad", test = "fisher_z", alpha = 0.05)

disco(tpcExample, my_pc, knowledge = kn)

# or using my_pc directly
my_pc <- my_pc |> set_knowledge(kn)
my_pc(tpcExample)
}
}
