% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_method.R
\name{pc}
\alias{pc}
\title{The Peter-Clark (PC) algorithm for causal discovery}
\usage{
pc(engine = c("tetrad", "pcalg", "bnlearn"), test, alpha = 0.05, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"tetrad"}}{Tetrad Java library.}
\item{\code{"pcalg"}}{\pkg{pcalg} R package.}
\item{\code{"bnlearn"}}{\pkg{bnlearn} R package.}
}}

\item{test}{Character; name of the conditional‐independence test.}

\item{alpha}{Numeric; significance level for the CI tests.}

\item{...}{Additional arguments passed to the chosen engine (e.g. test or algorithm parameters).}
}
\value{
A function of class \code{"fci"} that takes a single argument \code{data}
(a data frame) and returns an \code{discography} object.
}
\description{
Run the PC algorithm for causal discovery using one of several engines.
}
