% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcalg_search.R
\name{pcalgSearch}
\alias{pcalgSearch}
\title{Class for pcalg search algorithms}
\description{
This class implements the search algorithms from the pcalg package.
It allows to set the data, sufficient statistics, test, score, and algorithm.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pcalgSearch-new}{\code{pcalgSearch$new()}}
\item \href{#method-pcalgSearch-set_params}{\code{pcalgSearch$set_params()}}
\item \href{#method-pcalgSearch-set_data}{\code{pcalgSearch$set_data()}}
\item \href{#method-pcalgSearch-set_suff_stat}{\code{pcalgSearch$set_suff_stat()}}
\item \href{#method-pcalgSearch-set_test}{\code{pcalgSearch$set_test()}}
\item \href{#method-pcalgSearch-set_score}{\code{pcalgSearch$set_score()}}
\item \href{#method-pcalgSearch-set_alg}{\code{pcalgSearch$set_alg()}}
\item \href{#method-pcalgSearch-set_knowledge}{\code{pcalgSearch$set_knowledge()}}
\item \href{#method-pcalgSearch-run_search}{\code{pcalgSearch$run_search()}}
\item \href{#method-pcalgSearch-clone}{\code{pcalgSearch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pcalgSearch-new"></a>}}
\if{latex}{\out{\hypertarget{method-pcalgSearch-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pcalgSearch$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pcalgSearch-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-pcalgSearch-set_params}{}}}
\subsection{Method \code{set_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pcalgSearch$set_params(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pcalgSearch-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-pcalgSearch-set_data}{}}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pcalgSearch$set_data(data, set_suff_stat = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pcalgSearch-set_suff_stat"></a>}}
\if{latex}{\out{\hypertarget{method-pcalgSearch-set_suff_stat}{}}}
\subsection{Method \code{set_suff_stat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pcalgSearch$set_suff_stat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pcalgSearch-set_test"></a>}}
\if{latex}{\out{\hypertarget{method-pcalgSearch-set_test}{}}}
\subsection{Method \code{set_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pcalgSearch$set_test(method, alpha = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pcalgSearch-set_score"></a>}}
\if{latex}{\out{\hypertarget{method-pcalgSearch-set_score}{}}}
\subsection{Method \code{set_score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pcalgSearch$set_score(method, params = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pcalgSearch-set_alg"></a>}}
\if{latex}{\out{\hypertarget{method-pcalgSearch-set_alg}{}}}
\subsection{Method \code{set_alg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pcalgSearch$set_alg(method)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pcalgSearch-set_knowledge"></a>}}
\if{latex}{\out{\hypertarget{method-pcalgSearch-set_knowledge}{}}}
\subsection{Method \code{set_knowledge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pcalgSearch$set_knowledge(knowledge_obj, directed_as_undirected = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pcalgSearch-run_search"></a>}}
\if{latex}{\out{\hypertarget{method-pcalgSearch-run_search}{}}}
\subsection{Method \code{run_search()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pcalgSearch$run_search(data, set_suff_stat = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pcalgSearch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pcalgSearch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pcalgSearch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
