% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-caugi.R
\name{plot.knowledge}
\alias{plot.knowledge}
\title{Plot a Knowledge Object}
\usage{
\method{plot}{knowledge}(x, x_jitter = 0, vertex_size_scale = 1, ...)
}
\arguments{
\item{x}{A \code{knowledge} object.}

\item{x_jitter}{Amount of jitter to apply to x positions of tiered nodes (default 0).}

\item{vertex_size_scale}{Scaling factor for vertex sizes (default 1).}

\item{...}{Additional arguments passed to igraph \code{plot}.}
}
\value{
A plot of the knowledge structure.
}
\description{
Plot a knowledge object (tiers + required + forbidden)
}
\examples{
data(tpc_example)

kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

plot(kn)

}
