% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_caugi.R
\name{plot.knowledge}
\alias{plot.knowledge}
\title{Plot a Knowledge Object}
\usage{
\method{plot}{knowledge}(x, ...)
}
\arguments{
\item{x}{A \code{knowledge} object.}

\item{...}{Additional arguments passed to igraph \code{plot}.}
}
\value{
A plot of the knowledge structure.
}
\description{
Plot a knowledge object (tiers + required + forbidden)
}
\examples{
data("tpcExample")

kn <- knowledge(
  tpcExample,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

plot(kn)

}
