% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_caugi.R
\name{plot.knowledge}
\alias{plot.knowledge}
\title{Plot a Knowledge Object}
\usage{
\method{plot}{knowledge}(x, jitter = 0, space = 5, vertex_size_scale = 1, ...)
}
\arguments{
\item{x}{A \code{knowledge} object.}

\item{jitter}{Amount of horizontal jitter to apply to tiers (default 0).}

\item{...}{Additional arguments passed to igraph \code{plot}.}

\item{vertex_sizes}{Base size for vertices (default 1).}
}
\value{
A plot of the knowledge structure.
}
\description{
Plot a knowledge object (tiers + required + forbidden)
}
\examples{
data("tpc_example")

kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

plot(kn)

}
