% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_caugi.R
\name{plot.knowledgeable_caugi}
\alias{plot.knowledgeable_caugi}
\title{Plot a Causal Graph from a \code{knowledgeable_caugi} Object
This function visualizes a causal graph stored within a \code{caugi} object.}
\usage{
\method{plot}{knowledgeable_caugi}(x, ...)
}
\arguments{
\item{x}{A \code{caugi} object containing the causal graph and knowledge.}

\item{...}{Additional arguments passed to the plot function.}
}
\value{
A plot of the causal graph.
}
\description{
Plot a Causal Graph from a \code{knowledgeable_caugi} Object
This function visualizes a causal graph stored within a \code{caugi} object.
}
