% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tpdag}
\alias{plot.tpdag}
\title{Plot temporal partially directed acyclic graph (TPDAG)}
\usage{
\method{plot}{tpdag}(x, ...)
}
\arguments{
\item{x}{tpdag (temporal partially directed acyclic graph) object
to be plotted (as outputted from \code{\link{tpc}}).}

\item{...}{Further plotting arguments passed along to \code{\link{plotTempoMech}}.}
}
\value{
No return value, the function is called for its side-effects (plotting).
}
\description{
Plot temporal partially directed acyclic graph (TPDAG)
}
