% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTempoMech.R
\name{plotTempoMech}
\alias{plotTempoMech}
\title{Plot temporal data generating mechanism}
\usage{
plotTempoMech(
  x,
  addTimeAxis = TRUE,
  addPsi = TRUE,
  varLabels = NULL,
  periodLabels = NULL,
  colors = NULL,
  ...
)
}
\arguments{
\item{x}{The tpdag/tskeleton or tamat to plot.}

\item{addTimeAxis}{Logical indicating whether a time axis should be
added to the plot.}

\item{addPsi}{Logical indicating whether the sparsity level should be
added to the plot.}

\item{varLabels}{A named list of variable labels.}

\item{periodLabels}{A character vector with labels for periods.}

\item{colors}{A character vector with colors to use for marking periods.
Should have at least as many elements as the numbers of periods.}

\item{...}{Additional arguments passed to \code{\link[igraph]{plot.igraph}}.}
}
\value{
No return value, the function is called for its side-effects (plotting).
}
\description{
Plots tpdag, tskeleton and tamat objects.
}
