% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{+.knowledge}
\alias{+.knowledge}
\title{Merge two \code{knowledge} objects}
\usage{
\method{+}{knowledge}(.kn1, .kn2)
}
\arguments{
\item{.kn1}{A \code{knowledge} object.}

\item{.kn2}{Another \code{knowledge} object.}
}
\description{
Merge two \code{knowledge} objects
}
