% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{+.knowledge}
\alias{+.knowledge}
\title{Merge two \code{knowledge} objects}
\usage{
\method{+}{knowledge}(.kn1, .kn2)
}
\arguments{
\item{.kn1}{A \code{knowledge} object.}

\item{.kn2}{Another \code{knowledge} object.}
}
\description{
Merge two \code{knowledge} objects
}
\examples{
### +.knowledge example ###

# Create two knowledge objects
kn1 <- knowledge(
  tier(
    1 ~ V1,
    2 ~ V2
  ),
  required(V1 ~ V2)
)

kn2 <- knowledge(
  tier(3 ~ V3),
  forbidden(V2 ~ V3)
)

kn_merged <- kn1 + kn2

# Error paths
# Merging with conflicting tier information

kn1 <- knowledge(
  tier(
    1 ~ V1,
    2 ~ V2
  )
)

kn2 <- knowledge(
  tier(3 ~ V2)
)

try(kn1 + kn2)

kn2 <- knowledge(
  tier(1 ~ V1 + V2)
)

try(kn1 + kn2)

# Requried / forbidden violations

kn1 <- knowledge(
  forbidden(V1 ~ V2)
)

kn2 <- knowledge(
  required(V1 ~ V2)
)

try(kn1 + kn2)
}
