% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{precision}
\alias{precision}
\title{Precision}
\usage{
precision(confusion)
}
\arguments{
\item{confusion}{Confusion matrix as obtained from \link{confusion}}
}
\value{
A numeric in [0,1].
}
\description{
Computes precision (aka positive predictive value) from a confusion matrix,
see \link{confusion}.
Precision is defined as TP/(TP + FP), where TP are true positives and
FP are false positives. If TP + FP = 0, 0 is returned.
}
