% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{print.knowledge}
\alias{print.knowledge}
\title{Print a \code{knowledge} object}
\usage{
\method{print}{knowledge}(x, ...)
}
\arguments{
\item{x}{A \code{knowledge} object.}

\item{...}{Additional arguments (not used).}
}
\description{
Print a \code{knowledge} object
}
