% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{print.Knowledge}
\alias{print.Knowledge}
\title{Print a Knowledge Object}
\usage{
\method{print}{Knowledge}(x, ...)
}
\arguments{
\item{x}{A \code{Knowledge} object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the \code{Knowledge} object.
}
\description{
Print a Knowledge Object
}
\examples{
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)
print(kn)

}
