% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrad_rdata_utils.R
\name{rdata_to_tetrad}
\alias{rdata_to_tetrad}
\title{Tetrad R Data Utilities}
\usage{
rdata_to_tetrad(df)
}
\arguments{
\item{df}{A data frame to be converted to a Tetrad Java object.}
}
\value{
A Tetrad Java object representing the data frame.
}
\description{
\code{tetrad_rdata_utils} provides functions to convert between R data frames and Tetrad Java objects.
}
\details{
This function is made to be used internally with the TetradSearch class.
The function will copy the data into the Java heap, so be careful with larger data frames.
This function was provided by Joseph Ramsey, and slightly modified by Frederik Fabricius-Bjerre.
}
