% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{recall}
\alias{recall}
\title{Recall}
\usage{
recall(confusion)
}
\arguments{
\item{confusion}{Confusion matrix as obtained from \link{confusion}}
}
\value{
A numeric in [0,1].
}
\description{
Computes recall from a confusion matrix, see \link{confusion}.
Recall is defined as TP/(TP + FN), where TP are true positives and
FN are false negatives. If TP + FN = 0, 0 is returned.
}
