% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{shd}
\alias{hd}
\alias{aid}
\title{Objects exported from other packages}
\seealso{
Other metrics: 
\code{\link{confusion}()},
\code{\link{evaluate}()},
\code{\link{f1_score}()},
\code{\link{false_omission_rate}()},
\code{\link{fdr}()},
\code{\link{g1_score}()},
\code{\link{npv}()},
\code{\link{precision}()},
\code{\link{recall}()},
\code{\link{specificity}()}
}
\concept{metrics}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{caugi}{\code{\link[caugi]{aid}}, \code{\link[caugi]{hd}}, \code{\link[caugi]{shd}}}
}}

