% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-causaldisco-algs.R
\name{register_tetrad_algorithm}
\alias{register_tetrad_algorithm}
\title{Register a New Tetrad Algorithm}
\usage{
register_tetrad_algorithm(name, setup_fun)
}
\arguments{
\item{name}{Algorithm name (string)}

\item{setup_fun}{A function that sets up the Tetrad search object for the
algorithm. It should take the same arguments as the runner function for the algorithm.}
}
\description{
Registers a new Tetrad algorithm by adding it to the internal registry. The \code{setup_fun()} should be a function that
takes the same arguments as the runner function for the algorithm and sets up the Tetrad search object accordingly.
This allows you to extend the set of Tetrad algorithms that can be used with causalDisco.
}
\seealso{
Other Extending causalDisco: 
\code{\link{distribute_engine_args}()},
\code{\link{list_registered_tetrad_algorithms}()},
\code{\link{new_disco_method}()},
\code{\link{reset_tetrad_alg_registry}()}
}
\concept{Extending causalDisco}
\concept{extending_causalDisco}
