% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{remove_edges}
\alias{remove_edges}
\title{Remove edges from a knowledge object}
\usage{
remove_edges(.kn, ...)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{...}{One or more two‐sided formulas, e.g. \code{A ~ B} or
\code{starts_with("X") ~ Y}.}
}
\value{
The updated \code{knowledge} object.
}
\description{
Drop any directed edge(s) matching the two‐sided formulas you supply.
Errors if no edges matched.
}
\examples{
# remove variables and their incident edges
data(tpc_example)

kn <- knowledge(
  head(tpc_example),
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    oldage ~ starts_with("old")
  ),
  child_x1 \%-->\% youth_x3
)
print(kn)

kn <- remove_edges(kn, child_x1 ~ youth_x3)
print(kn)

kn <- remove_vars(kn, starts_with("child_"))
print(kn)

kn <- remove_tiers(kn, "child")
print(kn)
}
\seealso{
Other knowledge functions: 
\code{\link{+.knowledge}()},
\code{\link{add_exogenous}()},
\code{\link{add_tier}()},
\code{\link{add_to_tier}()},
\code{\link{add_vars}()},
\code{\link{as_bnlearn_knowledge}()},
\code{\link{as_pcalg_constraints}()},
\code{\link{as_tetrad_knowledge}()},
\code{\link{deparse_knowledge}()},
\code{\link{forbid_edge}()},
\code{\link{forbid_tier_violations}()},
\code{\link{get_tiers}()},
\code{\link{knowledge}()},
\code{\link{remove_tiers}()},
\code{\link{remove_vars}()},
\code{\link{reorder_tiers}()},
\code{\link{reposition_tier}()},
\code{\link{require_edge}()},
\code{\link{seq_tiers}()},
\code{\link{unfreeze}()}
}
\concept{knowledge}
\concept{knowledge functions}
