% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{remove_tiers}
\alias{remove_tiers}
\title{Remove entire tiers from a knowledge object}
\usage{
remove_tiers(.kn, ...)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{...}{Tier labels (unquoted or character) or numeric indices.}
}
\value{
An updated \code{knowledge} object.
}
\description{
Drops tier definitions (and un‐tiers any vars assigned to them).
}
\examples{
### remove_vars() example ###

# remove variables and their incident edges
data(tpcExample)

kn <- knowledge(
  head(tpcExample),
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    oldage ~ starts_with("old")
  ),
  required(child_x1 ~ youth_x3)
)
print(kn)

kn <- remove_edges(kn, child_x1 ~ youth_x3)
print(kn)

kn <- remove_vars(kn, starts_with("child_"))
print(kn)

kn <- remove_tiers(kn, "child")
print(kn)
}
