% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{remove_tiers}
\alias{remove_tiers}
\title{Remove entire tiers from a knowledge object}
\usage{
remove_tiers(.kn, ...)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{...}{Tier labels (unquoted or character) or numeric indices.}
}
\value{
An updated \code{knowledge} object.
}
\description{
Drops tier definitions (and un‐tiers any vars assigned to them).
}
