% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{remove_vars}
\alias{remove_vars}
\title{Remove variables (and their edges) from a knowledge object}
\usage{
remove_vars(.kn, ...)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{...}{Unquoted variable names or tidy‐select helpers.}
}
\value{
An updated \code{knowledge} object.
}
\description{
Drops the given variables from \code{kn$vars}, and automatically removes
any edges that mention them.
}
