% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{remove_vars}
\alias{remove_vars}
\title{Remove variables (and their edges) from a knowledge object}
\usage{
remove_vars(.kn, ...)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{...}{Unquoted variable names or tidy‐select helpers.}
}
\value{
An updated \code{knowledge} object.
}
\description{
Drops the given variables from \code{kn$vars}, and automatically removes
any edges that mention them.
}
\examples{
### remove_vars() example ###

# remove variables and their incident edges
data(tpcExample)

kn <- knowledge(
  head(tpcExample),
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    oldage ~ starts_with("old")
  ),
  required(child_x1 ~ youth_x3)
)
print(kn)

kn <- remove_edges(kn, child_x1 ~ youth_x3)
print(kn)

kn <- remove_vars(kn, starts_with("child_"))
print(kn)

kn <- remove_tiers(kn, "child")
print(kn)
}
