% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{require_edge}
\alias{require_edge}
\title{Add required edges}
\usage{
require_edge(.kn, ...)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{...}{One or more two-sided formulas.}
}
\value{
The updated \code{knowledge} object.
}
\description{
Require one or more directed edges.
Arguments follow the same rules as \strong{\code{forbid_edge()}} but a required edge
may only be given in \emph{one} direction (\code{X ~ Y} \strong{or} \code{Y ~ X}, not both).
}
