% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{require_edge}
\alias{require_edge}
\title{Add a forbidden edge to a knowledge object}
\usage{
require_edge(.kn, ...)
}
\arguments{
\item{.kn}{A \code{knowledge} object.}

\item{...}{Either a two-sided formula (\code{A ~ C}) \emph{or} \code{from}, \code{to}.}
}
\description{
These edges are not allowed to be present in the final graph.
}
