% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco-method.R
\name{set_knowledge}
\alias{set_knowledge}
\alias{set_knowledge.disco_method}
\title{Set Background Knowledge to Disco Method}
\usage{
set_knowledge(method, knowledge)

\method{set_knowledge}{disco_method}(method, knowledge)
}
\arguments{
\item{method}{A \code{"disco_method"} function.}

\item{knowledge}{A knowledge object appropriate for the engine.}
}
\description{
Set Background Knowledge to Disco Method
}
