% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{shd}
\alias{shd}
\title{Structural hamming distance between adjacency matrices}
\usage{
shd(est_amat, true_amat)
}
\arguments{
\item{est_amat}{Estimated adjacency matrix}

\item{true_amat}{True adjacency matrix}
}
\value{
A numeric (a non-negative integer).
}
\description{
Computes the structural hamming distance between two adjacency matrices. This implementation
is a modification of the \code{\link[pcalg]{shd}} function from the pcalg package, but here we
avoid working on the heavy \code{graphNEL} objects for representing graphs that are used in the
pcalg package.
}
\details{
Note that the function is symmetric in the two inputted adjacency matrices.
}
