% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-dag.R
\name{sim_dag}
\alias{sim_dag}
\title{Simulate a random DAG}
\usage{
sim_dag(p, sparsity = NULL, sparsity_lim = c(0, 0.8), permute = TRUE)
}
\arguments{
\item{p}{The number of nodes.}

\item{sparsity}{A numeric in [0,1] giving the proportion of possible
edges that should be removed. If \code{sparsity} is supplied, the value in
\code{sparsity_lim} is ignored.}

\item{sparsity_lim}{A numeric vector of length 2 giving the interval from
which a sparsity value is drawn when \code{sparsity = NULL}. Both values
must lie in [0,1].}

\item{permute}{If \code{FALSE}, the adjacency matrix will include nodes
in their causal ordering. This is avoided by setting \code{permute = TRUE},
in which case the node order is permuted randomly.}
}
\value{
An adjacency matrix.
}
\description{
Simulates a random directed acyclic graph adjacency (DAG) matrix with
the provided edge sparsity. The edge sparsity is the percentage of edges
that are absent, relative to a fully connected DAG.
}
\examples{
# Simulate a DAG adjacency matrix with 5 nodes and sparsity 0.5
sim_dag(5, sparsity = 0.5)

}
