% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(confusion)
}
\arguments{
\item{confusion}{Confusion matrix as obtained from \link{confusion}}
}
\value{
A numeric in [0,1].
}
\description{
Computes specificity from a confusion matrix, see \link{confusion}.
Specificity is defined as TN/(TN + FP), where TN are true negatives and
FP are false positives. If TN + FP = 0, 0 is returned.
}
