% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco-class.R
\name{summary.disco}
\alias{summary.disco}
\title{Summarize a disco Object}
\usage{
\method{summary}{disco}(object, ...)
}
\arguments{
\item{object}{A \code{disco} object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the \code{disco} object.
}
\description{
Summarize a disco Object
}
\examples{
data(tpc_example)
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)
cd_tges <- tpc(engine = "causalDisco", test = "fisher_z")
disco_cd_tges <- disco(data = tpc_example, method = cd_tges, knowledge = kn)
summary(disco_cd_tges)

}
