% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco-class.R
\name{summary.Disco}
\alias{summary.Disco}
\title{Summarize a Disco Object}
\usage{
\method{summary}{Disco}(object, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{summary()} for \code{Disco} objects is deprecated. Use \code{\link[=print]{print()}} instead.
}
\keyword{internal}
