% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{summary.Knowledge}
\alias{summary.Knowledge}
\title{Summarize a Knowledge Object}
\usage{
\method{summary}{Knowledge}(object, ...)
}
\arguments{
\item{object}{A \code{Knowledge} object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the \code{Knowledge} object.
}
\description{
Summarize a Knowledge Object
}
\examples{
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)
summary(kn)

}
