% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge.R
\name{summary.knowledge}
\alias{summary.knowledge}
\title{Summary method for knowledge objects}
\usage{
\method{summary}{knowledge}(object, ...)
}
\arguments{
\item{object}{A \code{knowledge} object.}

\item{...}{Additional arguments (not used).}
}
\description{
Summary method for knowledge objects
}
\examples{
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)
summary(kn)

}
