% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetrad_graph.R
\name{tetrad_graph}
\alias{tetrad_graph}
\title{Build a lightweight S3 wrapper around a Tetrad PAG string}
\usage{
tetrad_graph(x)
}
\arguments{
\item{x}{Character scalar in the form
"V1,V2,V3\n0,2,0\n3,0,1\n0,3,0\n"}
}
\value{
An object of class <tetrad_graph>
}
\description{
Build a lightweight S3 wrapper around a Tetrad PAG string
}
