% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfci.R
\name{tfci}
\alias{tfci}
\title{The Temporal Fast Causal Inference (FCI) algorithm for causal discovery}
\usage{
tfci(engine = c("causalDisco"), test, alpha = 0.05, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"causalDisco"}}{causalDisco library.}
}}

\item{test}{Character; name of the conditional‐independence test.}

\item{alpha}{Numeric; significance level for the CI tests.}

\item{...}{Additional arguments passed to the chosen engine
(e.g. test or algorithm parameters).}
}
\value{
A function of class \code{"tfci"} that takes a single argument \code{data}
(a data frame) and returns a \code{caugi} and \code{knowledge} (\code{knowledgeable_caugi})
object.
}
\description{
Run the temporal FCI algorithm for causal discovery using causalDisco.
}
\details{
For specific details on the supported scores, tests, and parameters for the engine, see:
\itemize{
\item \code{\link{CausalDiscoSearch}} for \pkg{causalDisco}.
}
}
\examples{
data("tpc_example")

kn <- knowledge(
  tpc_example,
  tier(
    child ~ tidyselect::starts_with("child"),
    youth ~ tidyselect::starts_with("youth"),
    oldage ~ tidyselect::starts_with("oldage")
  )
)

# Recommended path using disco()
my_tfci <- tfci(engine = "causalDisco", test = "fisher_z", alpha = 0.05)

disco(tpc_example, my_tfci, knowledge = kn)

# or using my_tfci directly
my_tfci <- my_tfci |> set_knowledge(kn)
my_tfci(tpc_example)

# Also possible: using tfci_run()
tfci_run(tpc_example, test = cor_test, knowledge = kn)
}
