% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tges}
\alias{tges}
\title{The Temporal GES algorithm for causal discovery}
\usage{
tges(engine = c("causalDisco"), score, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"causalDisco"}}{causalDisco library.}
}}

\item{score}{Character; name of the scoring function to use.}

\item{...}{Additional arguments passed to the chosen engine
(e.g. test or algorithm parameters).}
}
\value{
A function of class \code{"tges"} that takes a single argument \code{data}
(a data frame) and returns an \code{discography} object.
}
\description{
Run the Temporal GES algorithm for causal discovery using the causalDisco
engine.
}
