% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tges}
\alias{tges}
\title{TGES Algorithm for Causal Discovery}
\usage{
tges(engine = c("causalDisco"), score, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"causalDisco"}}{causalDisco library.}
}}

\item{score}{Character; name of the scoring function to use.}

\item{...}{Additional arguments passed to the chosen engine
(e.g. test or algorithm parameters).}
}
\value{
A function of class \code{"tges"} that takes a single argument \code{data}
(a data frame) and returns a \code{caugi} (of class "PDAG") and \code{knowledge} (\code{knowledgeable_caugi})
object.
}
\description{
Run the Temporal GES algorithm for causal discovery using the causalDisco
engine.
}
\details{
For specific details on the supported scores, see \code{\link{CausalDiscoSearch}}. For additional parameters passed
via \code{...}, see \code{\link[=tges_run]{tges_run()}}.
}
\examples{
# Recommended route using disco:
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

my_tges <- tges(engine = "causalDisco", score = "tbic")

disco(tpc_example, my_tges, knowledge = kn)

# another way to run it

my_tges <- my_tges |>
  set_knowledge(kn)
my_tges(tpc_example)


# or you can run directly with tges_run()

data(tpc_example)

score_bic <- new(
  "TemporalBIC",
  data = tpc_example,
  nodes = colnames(tpc_example),
  knowledge = kn
)

res_bic <- tges_run(score_bic)
res_bic
}
\seealso{
Other causal discovery algorithms: 
\code{\link{boss}()},
\code{\link{boss_fci}()},
\code{\link{fci}()},
\code{\link{ges}()},
\code{\link{gfci}()},
\code{\link{grasp}()},
\code{\link{grasp_fci}()},
\code{\link{pc}()},
\code{\link{sp_fci}()},
\code{\link{tfci}()},
\code{\link{tpc}()}
}
\concept{causal discovery algorithms}
\concept{cd_algorithms}
