% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges.R
\name{tges}
\alias{tges}
\title{The Temporal GES algorithm for causal discovery}
\usage{
tges(engine = c("causalDisco"), score, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"causalDisco"}}{causalDisco library.}
}}

\item{score}{Character; name of the scoring function to use.}

\item{...}{Additional arguments passed to the chosen engine
(e.g. test or algorithm parameters).}
}
\value{
A function of class \code{"tges"} that takes a single argument \code{data}
(a data frame) and returns a \code{caugi} and \code{knowledge} (\code{knowledgeable_caugi})
object.
}
\description{
Run the Temporal GES algorithm for causal discovery using the causalDisco
engine.
}
\details{
For specific details on the supported scores, tests, and parameters for the engine, see:
\itemize{
\item \code{\link{CausalDiscoSearch}} for \pkg{causalDisco}.
}
}
\examples{
# Recommended route using disco:
kn <- knowledge(
  tpc_example,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

my_tges <- tges(engine = "causalDisco", score = "tbic")

disco(tpc_example, my_tges, knowledge = kn)

# another way to run it

my_tges <- my_tges |>
  set_knowledge(kn)
my_tges(tpc_example)


# or you can run directly with tges_run()

data("tpc_example")

score_bic <- new(
  "TemporalBIC",
  data = tpc_example,
  nodes = colnames(tpc_example),
  knowledge = kn
)

res_bic <- tges_run(score_bic)
res_bic
}
