% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges_run.R
\name{tges_run}
\alias{tges_run}
\title{Estimate the restricted Markov equivalence class using Temporal Greedy Equivalence Search}
\usage{
tges_run(score, verbose = FALSE)
}
\arguments{
\item{score}{tiered scoring object to be used.
At the moment only scores supported are
\itemize{
\item \code{\linkS4class{TemporalBIC}} and
\item \code{\linkS4class{TemporalBDeu}}.
}}

\item{verbose}{indicates whether debug output should be printed.}
}
\value{
A \code{discography} object (the canonical representation for
\code{EssGraph}/\code{TEssGraph} results in this package).
}
\description{
Perform causal discovery using the temporal greedy equivalence search algorithm.
}
\author{
Tobias Ellegaard Larsen
}
