% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tges_run.R
\name{tges_run}
\alias{tges_run}
\title{Estimate the restricted Markov equivalence class using Temporal Greedy Equivalence Search}
\usage{
tges_run(score, verbose = FALSE)
}
\arguments{
\item{score}{tiered scoring object to be used.
At the moment only scores supported are
\itemize{
\item \code{\linkS4class{TemporalBIC}} and
\item \code{\linkS4class{TemporalBDeu}}.
}}

\item{verbose}{indicates whether debug output should be printed.}
}
\value{
A \code{caugi} and a \code{knowledge} (\code{knowledgeable_caugi}) object.
}
\description{
Perform causal discovery using the temporal greedy equivalence search algorithm.
}
\examples{
### tges_run() example ###

# Recommended route using disco:
kn <- knowledge(
  tpcExample,
  tier(
    child ~ starts_with("child"),
    youth ~ starts_with("youth"),
    old ~ starts_with("old")
  )
)

my_tges <- tges(engine = "causalDisco", score = "tbic")

disco(tpcExample, my_tges, knowledge = kn)

# another way to run it

my_tges <- my_tges |>
  set_knowledge(kn)
my_tges(tpcExample)


# or you can run directly with tges_run()

data("tpcExample")

score_bic <- new("TemporalBIC",
  data = tpcExample,
  nodes = colnames(tpcExample),
  knowledge = kn
)

res_bic <- tges_run(score_bic)
res_bic
}
\author{
Tobias Ellegaard Larsen
}
