% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpc.R
\name{tpc}
\alias{tpc}
\title{The Temporal Peter-Clark (PC) algorithm for causal discovery}
\usage{
tpc(engine = c("causalDisco"), test, alpha = 0.05, ...)
}
\arguments{
\item{engine}{Character; which engine to use. Must be one of:
\describe{
\item{\code{"causalDisco"}}{causalDisco library.}
}}

\item{test}{Character; name of the conditional‐independence test.}

\item{alpha}{Numeric; significance level for the CI tests.}

\item{...}{Additional arguments passed to the chosen engine
(e.g. test or algorithm parameters).}
}
\value{
A function of class \code{"tpc"} that takes a single argument \code{data}
(a data frame) and returns an \code{discography} object.
}
\description{
Run the temporal PC algorithm for causal discovery using causalDisco.
}
\examples{
### tpc() example ###

# Load data
data(tpcExample)

# Build knowledge
kn <- knowledge(
  tpcExample,
  tier(
    child ~ tidyselect::starts_with("child"),
    youth ~ tidyselect::starts_with("youth"),
    old ~ tidyselect::starts_with("old")
  )
)

# Recommended route using disco
my_tpc <- tpc(engine = "causalDisco", test = "fisher_z", alpha = 0.05)

disco(tpcExample, my_tpc, knowledge = kn)

# or using my_tpc directly

my_tpc <- my_tpc |> set_knowledge(kn)
my_tpc(tpcExample)

# Using tpc_run() directly

tpc_run(tpcExample, knowledge = kn, alpha = 0.01)

# Deprecated: using order prefixes (will warn)
testthat::expect_warning(
  tpc_run(tpcExample, order = c("child", "youth", "oldage"), alpha = 0.01)
)
}
