% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpc.R
\name{tpc}
\alias{tpc}
\title{Perform causal discovery using the temporal PC algorithm (TPC)}
\usage{
tpc(data, order, sparsity, test = NULL, ...)
}
\arguments{
\item{data}{A data.frame}

\item{order}{A character vector...}

\item{sparsity}{A numeric vector...}

\item{test}{A procedure for testing conditional independence...}

\item{...}{Further optional arguments which will be passed to the
testing procedure.}
}
\value{
Value blablabla tpdag
}
\description{
More description here.
}
